feat(landing): modular hero animation with realistic rocket launch
Split the monolithic hero-svg.tsx into eight focused files (constants,
defs, scenes for cargo+rocket, conveyor, animated layers, cubes, pipes,
plus an orchestrating index) so the animation surface is finally
navigable. Layer in a sequenced rocket launch: countdown ticks ->
BUILDING cube turns green on READY -> ignition burst with halo
overshoot -> pre-liftoff hold -> ascent -> trail fade. RE-LAUNCH now
plays a retro-thrust descent with dimmed flames, a two-tier ember
field (tight core sparks plus wider outer scatter), and a clip path
anchored at the pad's top rim so the flame never bleeds into or below
the cylinder. Adds a CSS-keyframe float on the isometric RE-LAUNCH
arrow with a prefers-reduced-motion fallback. CTA entrance and hover
transitions get shorter ease-out-quart curves that respect reduced
motion.
