+++ components/landing/svg/hero/scene-animated-layers.tsx
@@ -1,0 +1,239 @@
+const SceneAnimatedLayers = () => (
+  <>
+        <g className="layer1" style={{ visibility: "hidden" }}>
+          <rect
+            width="73.1803"
+            height="73.1803"
+            transform="matrix(0.866025 0.5 -0.866025 0.5 87.4678 45.0469)"
+            fill="white"
+            stroke="black"
+          />
+          <path
+            d="M89.3104 112.362C88.8623 112.621 88.3612 112.793 87.8069 112.878C87.2573 112.961 86.6915 112.949 86.1095 112.842C85.5322 112.733 84.9756 112.524 84.4398 112.214C83.9179 111.913 83.5599 111.597 83.366 111.266C83.1673 110.933 83.1373 110.604 83.2759 110.278C83.4145 109.953 83.7193 109.654 84.1904 109.382C84.8324 109.012 85.4906 108.816 86.1649 108.794C86.8393 108.768 87.5206 108.894 88.2088 109.174L87.1557 109.83C86.7908 109.684 86.4305 109.614 86.0749 109.622C85.7146 109.628 85.3821 109.718 85.0772 109.894C84.8139 110.046 84.6592 110.216 84.613 110.402C84.5668 110.589 84.6223 110.784 84.7793 110.986C84.9363 111.189 85.1858 111.389 85.5276 111.586C85.874 111.786 86.2227 111.932 86.5737 112.022C86.9247 112.108 87.2596 112.136 87.5783 112.106C87.897 112.077 88.1834 111.989 88.4374 111.842C88.77 111.65 88.9339 111.449 88.9293 111.238C88.9247 111.022 88.7792 110.805 88.4928 110.586L89.636 109.974C89.9732 110.244 90.1879 110.514 90.2803 110.786C90.3727 111.058 90.3403 111.328 90.1833 111.594C90.0309 111.858 89.7399 112.114 89.3104 112.362ZM91.618 110.902L88.4726 107.038L89.7197 106.318L96.4123 108.134L95.3176 108.766L93.9112 108.37L92.0198 109.462L92.7057 110.274L91.618 110.902ZM91.4517 108.79L92.7542 108.038L90.2046 107.318L91.4517 108.79ZM96.884 107.862L91.9649 105.022L93.1289 104.35L98.4774 105.182L95.0826 103.222L96.1357 102.614L101.055 105.454L99.8769 106.134L94.66 105.362L97.937 107.254L96.884 107.862ZM103.297 104.159L96.6041 102.343L97.6919 101.715L102.736 103.115L100.311 100.203L101.398 99.5753L104.537 103.443L103.297 104.159ZM105.786 102.722L102.64 98.8583L103.887 98.1383L110.58 99.9543L109.485 100.586L108.079 100.19L106.187 101.282L106.873 102.094L105.786 102.722ZM105.619 100.61L106.922 99.8583L104.372 99.1383L105.619 100.61ZM112.994 98.6884C112.569 98.9338 112.133 99.1031 111.685 99.1964C111.237 99.2898 110.798 99.3111 110.368 99.2604C109.939 99.2098 109.537 99.0898 109.163 98.9004L110.174 98.2604C110.396 98.3564 110.618 98.4204 110.84 98.4524C111.061 98.4791 111.278 98.4711 111.491 98.4284C111.708 98.3831 111.923 98.2991 112.135 98.1764C112.311 98.0751 112.438 97.9778 112.516 97.8844C112.599 97.7884 112.632 97.6978 112.613 97.6124C112.59 97.5244 112.516 97.4444 112.391 97.3724C112.276 97.3058 112.149 97.2644 112.01 97.2484C111.867 97.2298 111.683 97.2458 111.456 97.2964C111.225 97.3444 110.923 97.4364 110.549 97.5724C110.031 97.7591 109.583 97.8924 109.205 97.9724C108.83 98.0498 108.496 98.0698 108.2 98.0324C107.9 97.9924 107.611 97.8924 107.334 97.7324C107.038 97.5618 106.851 97.3711 106.773 97.1604C106.699 96.9471 106.738 96.7244 106.891 96.4924C107.043 96.2604 107.311 96.0338 107.694 95.8124C108.101 95.5778 108.516 95.4204 108.941 95.3404C109.366 95.2604 109.782 95.2471 110.188 95.3004C110.599 95.3511 110.983 95.4604 111.338 95.6284L110.341 96.2684C110.156 96.1778 109.962 96.1191 109.759 96.0924C109.551 96.0631 109.345 96.0671 109.142 96.1044C108.939 96.1418 108.743 96.2151 108.553 96.3244C108.313 96.4631 108.177 96.6031 108.145 96.7444C108.117 96.8831 108.188 97.0018 108.359 97.1004C108.479 97.1698 108.611 97.2111 108.754 97.2244C108.897 97.2324 109.071 97.2124 109.274 97.1644C109.482 97.1138 109.747 97.0271 110.071 96.9044C110.62 96.6938 111.098 96.5538 111.505 96.4844C111.911 96.4098 112.267 96.3964 112.572 96.4444C112.876 96.4924 113.154 96.5884 113.403 96.7324C113.708 96.9084 113.895 97.1071 113.964 97.3284C114.033 97.5444 113.985 97.7698 113.819 98.0044C113.657 98.2364 113.382 98.4644 112.994 98.6884Z"
+            fill="black"
+          />
+        </g>
+        <path
+          d="M246.285 212.054L284.106 218.844L310.702 223.515V236.842L326.826 246.603L327.99 208.797L417.335 216.315L460.809 216.428C461.287 213.537 458.856 211.393 457.581 210.683C428.86 194.276 369.596 160.293 362.307 155.621C355.018 150.95 347.729 153.303 344.996 155.064L246.285 212.054Z"
+          fill="white"
+        />
+        <path
+          d="M461.363 216.428C461.84 213.537 459.41 211.393 458.135 210.683C429.414 194.276 370.15 160.293 362.86 155.621C355.571 150.95 348.282 153.303 345.549 155.064L246.839 212.054C245.404 212.874 243.078 215.115 245.255 217.52"
+          stroke="black"
+        />
+        <path
+          d="M361.383 160.867C363.118 160.724 368.804 160.78 371.184 162.154L409.109 184.05C415.46 188.938 417.068 195.377 416.601 200.823L416.601 223.119C416.232 226.888 414.427 230.376 413.571 231.649C397.221 248.394 378.996 249.913 371.928 248.579C361.109 247.596 358.001 245.368 357.799 244.377V223.119L361.383 160.867Z"
+          fill="white"
+        />
+        <path
+          className="piston"
+          d="M413.571 209.81C409.109 216.679 389.031 229.13 371.928 226.983M413.571 209.81L413.571 231.649M413.571 209.81C414.983 207.83 416.28 204.56 416.601 200.823M371.928 226.983C358.245 225.266 358.79 223.405 357.799 223.119M371.928 226.983V248.579M357.799 223.119L361.383 160.867C363.118 160.724 368.804 160.78 371.184 162.154C373.564 163.527 396.963 177.037 409.109 184.05C415.46 188.938 417.068 195.377 416.601 200.823M357.799 223.119V244.377C358.001 245.368 361.109 247.596 371.928 248.579M371.928 248.579C378.996 249.913 397.221 248.394 413.571 231.649M413.571 231.649C414.427 230.376 416.232 226.888 416.601 223.119L416.601 200.823"
+          stroke="black"
+        />
+        <path
+          d="M288.64 196.07L311.692 182.761L312.186 204.501L310.205 239.325L290.127 228.074C286.094 225.814 284.967 223.504 284.907 222.632L284.907 200.711C284.903 200.622 284.903 200.534 284.907 200.446C284.977 198.898 286.268 197.439 288.64 196.07Z"
+          fill="white"
+        />
+        <path
+          className="piston"
+          d="M311.692 182.761L288.64 196.07C286.268 197.439 284.977 198.898 284.907 200.446M311.692 182.761L310.205 217.966M311.692 182.761L312.186 204.501L310.205 239.325M290.127 206.374C297.266 210.496 306.487 215.819 310.205 217.966M290.127 206.374V228.074M290.127 206.374C289.588 206.063 289.093 205.755 288.64 205.449C286.031 203.689 284.836 202.021 284.907 200.446M310.205 217.966V239.325M310.205 239.325L290.127 228.074M290.127 228.074C286.094 225.814 284.967 223.504 284.907 222.632L284.907 200.446"
+          stroke="black"
+          strokeLinejoin="round"
+        />
+        <path
+          d="M331.028 170.74L352.593 158.289C355.262 156.748 359.654 156.502 361.517 156.572C361.517 170.734 360.331 191.054 359.081 208.069L357.896 244.931L356.311 260.987C355.565 264.282 349.556 262.36 346.644 260.987L327.31 249.315V227.841L331.028 170.74Z"
+          fill="white"
+        />
+
+        <path
+          className="piston"
+          d="M327.31 227.841L331.028 170.74L352.593 158.289C355.262 156.748 359.654 156.502 361.517 156.572C361.517 170.734 360.331 191.054 359.081 208.069M327.31 227.841L346.644 239.004M327.31 227.841V249.315L346.644 260.987M346.644 239.004C348.131 239.863 355.901 242.038 356.311 239.433M346.644 239.004V260.987M356.311 239.433V260.987M356.311 239.433C356.671 237.145 357.905 224.068 359.081 208.069M356.311 260.987C355.565 264.282 349.556 262.36 346.644 260.987M356.311 260.987L357.896 244.931L359.081 208.069"
+          stroke="black"
+        />
+        <path
+          d="M247.445 212.235L284.381 218.162C284.381 219.267 284.165 221.686 289.038 224.61C293.911 227.534 310.216 236.829 310.634 236.829L326.804 246.445L346.39 257.878C349.076 259.221 354.936 261.345 356.886 259.096L358.745 242.87C360.951 243.788 366.166 245.267 371.649 245.855C378.342 246.573 395.489 247.067 413.831 229.199C415.136 227.53 417.71 222.235 417.194 215.97L460.991 215.501C461.148 215.875 460.691 217.478 457.832 219.125C454.973 220.773 392.47 256.419 359.364 275.89C357.688 277.034 352.638 279.163 346.119 276.415L247.308 219.116C245.716 218.135 242.642 214.583 247.445 212.235Z"
+          fill="white"
+        />
+        <path
+          d="M284.791 218.063C284.791 219.168 284.212 221.423 289.085 224.346C293.958 227.27 310.451 236.396 310.87 236.396M326.804 246.005L346.39 257.878C349.076 259.221 354.745 260.762 356.694 258.513L358.421 242.297C360.627 243.214 367.249 245.061 372.621 245.442C379.335 245.919 392.767 246.646 413.455 229.114C417.169 225.114 416.76 223.005 416.76 216.298"
+          stroke="black"
+        />
+        <path
+          d="M345.972 276.227C349.798 278.436 356.002 278.436 359.828 276.227V325.227C356.002 327.436 349.798 327.436 345.972 325.227V276.227Z"
+          fill="white"
+          stroke="black"
+        />
+        <rect
+          width="114"
+          height="49"
+          transform="matrix(0.866025 0.5 0 1 247.244 219.227)"
+          fill="white"
+          stroke="black"
+        />
+        <path
+          d="M244.375 215.227C244.375 216.789 245.472 218.203 247.245 219.227V268.227C245.472 267.203 244.375 265.789 244.375 264.227V215.227Z"
+          fill="white"
+          stroke="black"
+        />
+        <path
+          d="M461.424 215.227C461.424 216.789 460.328 218.203 458.555 219.227V268.227C460.328 267.203 461.424 265.789 461.424 264.227V215.227Z"
+          fill="white"
+          stroke="black"
+        />
+        <g className="layer2" style={{ visibility: "hidden" }}>
+          <rect
+            width="73.1803"
+            height="73.1803"
+            transform="matrix(0.866025 0.5 -0.866025 0.5 88.959 31.1172)"
+            fill="white"
+            stroke="black"
+          />
+          <path
+            d="M152.334 67.7077L25.5824 67.7077M88.9593 31.1174L88.9593 104.297M120.646 49.4126L57.2709 86.0026M110.084 43.3143L46.7083 79.9042M131.209 55.511L67.8335 92.1009M57.2712 49.4121L120.647 86.0021M46.7086 55.5105L110.084 92.1004M67.8338 43.3138L131.209 79.9037M115.365 82.9534C100.781 91.3734 77.1361 91.3734 62.5522 82.9534C47.9683 74.5334 47.9683 60.8818 62.5522 52.4618C77.1361 44.0418 100.781 44.0418 115.365 52.4618C129.949 60.8818 129.949 74.5334 115.365 82.9534ZM75.7557 75.3302C68.4637 71.1202 68.4637 64.2944 75.7557 60.0843C83.0477 55.8743 94.8703 55.8743 102.162 60.0843C109.454 64.2944 109.454 71.1202 102.162 75.3302C94.8703 79.5402 83.0477 79.5402 75.7557 75.3302ZM54.1024 75.0254L76.2837 87.8317C80.7204 90.3933 82.9388 91.6741 85.4969 92.154C87.7471 92.5761 90.1709 92.5761 92.4211 92.154C94.9791 91.6741 97.1975 90.3933 101.634 87.8317L123.816 75.0254C128.252 72.4638 130.471 71.183 131.302 69.7061C132.033 68.407 132.033 67.0076 131.302 65.7084C130.471 64.2315 128.252 62.9508 123.816 60.3892L101.634 47.5828C97.1975 45.0212 94.9791 43.7404 92.4211 43.2606C90.1709 42.8385 87.7471 42.8385 85.4969 43.2606C82.9388 43.7404 80.7204 45.0212 76.2837 47.5828L54.1024 60.3892C49.6656 62.9508 47.4472 64.2315 46.616 65.7084C45.8849 67.0076 45.8849 68.407 46.6161 69.7061C47.4472 71.183 49.6656 72.4638 54.1024 75.0254ZM81.565 47.5828L54.1024 63.4383C49.6656 65.9999 47.4472 67.2807 46.6161 68.7576C45.8849 70.0567 45.8849 71.4562 46.616 72.7553C47.4472 74.2322 49.6656 75.513 54.1024 78.0745L71.0023 87.8317C75.4391 90.3933 77.6575 91.6741 80.2156 92.154C82.4657 92.5761 84.8896 92.5761 87.1397 92.154C89.6978 91.6741 91.9162 90.3933 96.353 87.8317L123.816 71.9762C128.252 69.4146 130.471 68.1338 131.302 66.6569C132.033 65.3578 132.033 63.9584 131.302 62.6593C130.471 61.1824 128.252 59.9016 123.816 57.34L106.916 47.5828C102.479 45.0212 100.26 43.7404 97.7024 43.2606C95.4522 42.8385 93.0284 42.8385 90.7782 43.2606C88.2202 43.7404 86.0018 45.0212 81.565 47.5828ZM123.816 63.4383L96.353 47.5828C91.9162 45.0212 89.6978 43.7404 87.1397 43.2606C84.8896 42.8385 82.4658 42.8385 80.2156 43.2606C77.6575 43.7404 75.4391 45.0212 71.0023 47.5828L54.1024 57.34C49.6656 59.9016 47.4472 61.1824 46.6161 62.6593C45.8849 63.9584 45.8849 65.3578 46.6161 66.6569C47.4472 68.1338 49.6656 69.4146 54.1024 71.9762L81.565 87.8317C86.0018 90.3933 88.2202 91.6741 90.7782 92.154C93.0284 92.5761 95.4522 92.5761 97.7024 92.154C100.26 91.6741 102.479 90.3933 106.916 87.8317L123.816 78.0746C128.252 75.513 130.471 74.2322 131.302 72.7553C132.033 71.4562 132.033 70.0567 131.302 68.7576C130.471 67.2807 128.252 65.9999 123.816 63.4383Z"
+            stroke="#CACACA"
+            strokeWidth="0.5"
+          />
+          <path
+            d="M90.7669 98.4527C90.2912 98.7273 89.7762 98.8993 89.2219 98.9687C88.6677 99.038 88.1065 99.0153 87.5384 98.9007C86.9703 98.7807 86.4345 98.5753 85.931 98.2847C85.4368 97.9993 85.0881 97.6913 84.8849 97.3607C84.6817 97.03 84.6424 96.7007 84.7671 96.3727C84.8918 96.0393 85.199 95.7313 85.6886 95.4487C86.1135 95.2033 86.543 95.038 86.9772 94.9527C87.4114 94.862 87.8455 94.8407 88.2797 94.8887C88.7139 94.9367 89.1411 95.042 89.5614 95.2047L88.5222 95.8607C88.1989 95.7327 87.8756 95.6713 87.5522 95.6767C87.2289 95.6767 86.9056 95.77 86.5823 95.9567C86.3005 96.1193 86.1389 96.298 86.0973 96.4927C86.0604 96.6847 86.1227 96.8807 86.2844 97.0807C86.4507 97.278 86.6954 97.47 87.0188 97.6567C87.3467 97.846 87.6816 97.9887 88.0234 98.0847C88.3651 98.1753 88.7 98.2087 89.0279 98.1847C89.3605 98.158 89.6746 98.0593 89.9702 97.8887C90.1873 97.7633 90.3304 97.63 90.3997 97.4887C90.4736 97.3447 90.4782 97.2033 90.4136 97.0647C90.3535 96.9233 90.2358 96.7913 90.0602 96.6687L88.9379 97.3167L88.1688 96.8727L90.3235 95.6287L92.88 97.1047L92.2288 97.4807L91.1203 96.9047L91.3004 96.8647C91.4713 97.0167 91.5614 97.186 91.5706 97.3727C91.5798 97.5593 91.5175 97.7473 91.3835 97.9367C91.2496 98.1207 91.0441 98.2927 90.7669 98.4527ZM93.7203 96.6195L88.8013 93.7795L90.9559 92.5355C91.3162 92.3275 91.688 92.1822 92.0714 92.0995C92.4594 92.0142 92.8358 91.9942 93.2007 92.0395C93.5609 92.0822 93.8912 92.1902 94.1914 92.3635C94.4177 92.4942 94.5701 92.6382 94.6487 92.7955C94.7318 92.9502 94.7457 93.1075 94.6902 93.2675C94.6394 93.4249 94.5216 93.5729 94.3369 93.7115L94.2538 93.6875C94.5817 93.4982 94.9027 93.3955 95.2168 93.3795C95.5262 93.3609 95.838 93.4289 96.1521 93.5835L97.6901 94.3275L96.6163 94.9475L95.2791 94.2795C95.0759 94.1782 94.8773 94.1329 94.6833 94.1435C94.4893 94.1489 94.2676 94.2235 94.0182 94.3675L92.972 94.9715L94.7734 96.0115L93.7203 96.6195ZM92.0922 94.4635L93.076 93.8955C93.33 93.7489 93.4686 93.6022 93.4917 93.4555C93.5147 93.3035 93.4178 93.1649 93.2007 93.0395C92.979 92.9115 92.7411 92.8569 92.4871 92.8755C92.2284 92.8915 91.9628 92.9782 91.6903 93.1355L90.7412 93.6835L92.0922 94.4635ZM98.5511 93.8305L93.6321 90.9905L94.6851 90.3825L99.6042 93.2225L98.5511 93.8305ZM100.533 92.6859L95.6144 89.8459L97.3119 88.8659C98.1109 88.4046 98.94 88.1739 99.7991 88.1739C100.654 88.1713 101.476 88.3979 102.266 88.8539C103.051 89.3073 103.443 89.7793 103.443 90.2699C103.448 90.7579 103.058 91.2286 102.272 91.6819L100.533 92.6859ZM100.7 91.5659L101.344 91.1939C101.82 90.9193 102.044 90.6433 102.016 90.3659C101.988 90.0833 101.707 89.7873 101.171 89.4779C100.635 89.1686 100.125 89.0073 99.6397 88.9939C99.1548 88.9753 98.6744 89.1033 98.1987 89.3779L97.5543 89.7499L100.7 91.5659Z"
+            fill="black"
+          />
+        </g>
+
+        <g className="layer3" style={{ visibility: "hidden" }}>
+          <rect
+            width="73.1803"
+            height="73.1803"
+            transform="matrix(0.866025 0.5 -0.866025 0.5 88.96 16.6064)"
+            fill="white"
+            stroke="black"
+          />
+          <path
+            d="M88.6063 85.0619L83.6873 82.2219L85.3847 81.2419C86.1837 80.7806 87.0128 80.5499 87.8719 80.5499C88.7264 80.5472 89.5485 80.7739 90.3384 81.2299C91.1235 81.6832 91.5161 82.1552 91.5161 82.6459C91.5208 83.1339 91.1305 83.6046 90.3453 84.0579L88.6063 85.0619ZM88.7726 83.9419L89.4169 83.5699C89.8926 83.2952 90.1166 83.0192 90.0889 82.7419C90.0612 82.4592 89.7795 82.1632 89.2437 81.8539C88.7079 81.5446 88.1975 81.3832 87.7126 81.3699C87.2276 81.3512 86.7472 81.4792 86.2715 81.7539L85.6272 82.1259L88.7726 83.9419ZM93.4845 82.2455L88.5654 79.4055L91.988 77.4295L92.8748 77.9415L90.5053 79.3095L91.6346 79.9615L93.9209 78.6415L94.7939 79.1455L92.5076 80.4655L93.6507 81.1255L96.0756 79.7255L96.9624 80.2375L93.4845 82.2455ZM99.852 78.6972C99.427 78.9426 98.9906 79.1119 98.5425 79.2052C98.0945 79.2986 97.6557 79.3199 97.2262 79.2692C96.7966 79.2186 96.3948 79.0986 96.0207 78.9092L97.0322 78.2692C97.2539 78.3652 97.4756 78.4292 97.6973 78.4612C97.919 78.4879 98.1361 78.4799 98.3485 78.4372C98.5656 78.3919 98.7804 78.3079 98.9929 78.1852C99.1684 78.0839 99.2954 77.9866 99.3739 77.8932C99.4571 77.7972 99.4894 77.7066 99.4709 77.6212C99.4478 77.5332 99.3739 77.4532 99.2492 77.3812C99.1337 77.3146 99.0067 77.2732 98.8682 77.2572C98.725 77.2386 98.5402 77.2546 98.3139 77.3052C98.083 77.3532 97.7804 77.4452 97.4063 77.5812C96.889 77.7679 96.441 77.9012 96.0622 77.9812C95.6881 78.0586 95.3533 78.0786 95.0576 78.0412C94.7574 78.0012 94.4688 77.9012 94.1916 77.7412C93.896 77.5706 93.709 77.3799 93.6304 77.1692C93.5565 76.9559 93.5958 76.7332 93.7482 76.5012C93.9006 76.2692 94.1685 76.0426 94.5519 75.8212C94.9583 75.5866 95.374 75.4292 95.799 75.3492C96.2239 75.2692 96.6396 75.2559 97.046 75.3092C97.4571 75.3599 97.8405 75.4692 98.1961 75.6372L97.1985 76.2772C97.0137 76.1866 96.8197 76.1279 96.6165 76.1012C96.4086 76.0719 96.2031 76.0759 95.9999 76.1132C95.7967 76.1506 95.6004 76.2239 95.411 76.3332C95.1708 76.4719 95.0346 76.6119 95.0022 76.7532C94.9745 76.8919 95.0461 77.0106 95.217 77.1092C95.3371 77.1786 95.4687 77.2199 95.6119 77.2332C95.7551 77.2412 95.9283 77.2212 96.1315 77.1732C96.3394 77.1226 96.6049 77.0359 96.9283 76.9132C97.4779 76.7026 97.9559 76.5626 98.3624 76.4932C98.7689 76.4186 99.1245 76.4052 99.4293 76.4532C99.7342 76.5012 100.011 76.5972 100.261 76.7412C100.566 76.9172 100.753 77.1159 100.822 77.3372C100.891 77.5532 100.843 77.7786 100.676 78.0132C100.515 78.2452 100.24 78.4732 99.852 78.6972ZM102.402 77.0971L97.4828 74.2571L98.5359 73.6491L103.455 76.4891L102.402 77.0971ZM106.532 74.8403C106.057 75.115 105.542 75.287 104.987 75.3563C104.433 75.4257 103.872 75.403 103.304 75.2883C102.736 75.1683 102.2 74.963 101.696 74.6723C101.202 74.387 100.853 74.079 100.65 73.7483C100.447 73.4177 100.408 73.0883 100.532 72.7603C100.657 72.427 100.964 72.119 101.454 71.8363C101.879 71.591 102.308 71.4257 102.743 71.3403C103.177 71.2497 103.611 71.2283 104.045 71.2763C104.479 71.3243 104.906 71.4297 105.327 71.5923L104.288 72.2483C103.964 72.1203 103.641 72.059 103.318 72.0643C102.994 72.0643 102.671 72.1577 102.348 72.3443C102.066 72.507 101.904 72.6857 101.863 72.8803C101.826 73.0723 101.888 73.2683 102.05 73.4683C102.216 73.6657 102.461 73.8577 102.784 74.0443C103.112 74.2337 103.447 74.3763 103.789 74.4723C104.13 74.563 104.465 74.5963 104.793 74.5723C105.126 74.5457 105.44 74.447 105.736 74.2763C105.953 74.151 106.096 74.0177 106.165 73.8763C106.239 73.7323 106.244 73.591 106.179 73.4523C106.119 73.311 106.001 73.179 105.826 73.0563L104.703 73.7043L103.934 73.2603L106.089 72.0163L108.645 73.4923L107.994 73.8683L106.886 73.2923L107.066 73.2523C107.237 73.4043 107.327 73.5737 107.336 73.7603C107.345 73.947 107.283 74.135 107.149 74.3243C107.015 74.5083 106.809 74.6803 106.532 74.8403ZM109.486 73.0072L104.567 70.1672L105.731 69.4952L111.079 70.3272L107.684 68.3672L108.737 67.7592L113.656 70.5992L112.479 71.2792L107.262 70.5072L110.539 72.3992L109.486 73.0072Z"
+            fill="black"
+          />
+          <path
+            d="M114.687 45.3896L105.348 50.7818C104.257 51.4118 103.644 52.2663 103.644 53.1572C103.644 54.0482 104.257 54.9027 105.348 55.5327L114.687 60.9249"
+            stroke="#CDCDCD"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+          <path
+            d="M60.7891 48.8096L64.204 50.7812C65.8408 51.7262 68.0608 52.2571 70.3756 52.2571C72.6903 52.2571 74.9103 51.7262 76.5471 50.7812C77.6383 50.1512 79.1183 49.7972 80.6615 49.7972C82.2047 49.7972 83.6847 50.1512 84.7759 50.7812C87.0388 52.0877 90.7417 52.0877 93.0046 50.7812C94.0958 50.1512 94.7088 49.2967 94.7088 48.4057C94.7088 47.5148 94.0958 46.6603 93.0046 46.0303C90.7417 44.7238 90.7417 42.5859 93.0046 41.2794L99.5259 37.5144"
+            stroke="#CDCDCD"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+          <path
+            d="M107.303 66.1625L99.1774 61.4711C98.0862 60.8411 96.6062 60.4871 95.063 60.4871C93.5198 60.4871 92.0398 60.8411 90.9486 61.4711C89.8574 62.1011 88.3774 62.455 86.8342 62.455C85.291 62.455 83.8111 62.1011 82.7199 61.4711L80.6627 60.2833C79.5715 59.6533 78.0915 59.2994 76.5483 59.2994C75.0051 59.2994 73.5251 59.6533 72.4339 60.2833L66.3652 63.7871"
+            stroke="#CDCDCD"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+          <path
+            d="M109.422 65.0568C120.806 58.5097 120.841 47.8745 109.501 41.3025C98.1608 34.7304 79.7402 34.7102 68.3571 41.2573C56.9739 47.8044 56.9389 58.4395 68.2788 65.0116C79.6186 71.5836 98.0393 71.6039 109.422 65.0568Z"
+            stroke="#CDCDCD"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+        </g>
+        <g className="layer4" style={{ visibility: "hidden" }}>
+          <rect
+            width="73.1803"
+            height="73.1803"
+            transform="matrix(0.866025 0.5 -0.866025 0.5 88.8896 0.5)"
+            fill="white"
+            stroke="black"
+          />
+          <path
+            d="M90.5946 67.8955C90.1697 68.1408 89.7332 68.3101 89.2852 68.4035C88.8372 68.4968 88.3984 68.5181 87.9689 68.4675C87.5393 68.4168 87.1375 68.2968 86.7633 68.1075L87.7749 67.4675C87.9966 67.5635 88.2183 67.6275 88.44 67.6595C88.6617 67.6861 88.8788 67.6781 89.0912 67.6355C89.3083 67.5901 89.5231 67.5061 89.7355 67.3835C89.9111 67.2821 90.0381 67.1848 90.1166 67.0915C90.1997 66.9955 90.2321 66.9048 90.2136 66.8195C90.1905 66.7315 90.1166 66.6515 89.9919 66.5795C89.8764 66.5128 89.7494 66.4715 89.6108 66.4555C89.4677 66.4368 89.2829 66.4528 89.0566 66.5035C88.8256 66.5515 88.5231 66.6435 88.149 66.7795C87.6317 66.9661 87.1837 67.0995 86.8049 67.1795C86.4308 67.2568 86.0959 67.2768 85.8003 67.2395C85.5001 67.1995 85.2114 67.0995 84.9343 66.9395C84.6387 66.7688 84.4516 66.5781 84.3731 66.3675C84.2992 66.1541 84.3385 65.9315 84.4909 65.6995C84.6433 65.4675 84.9112 65.2408 85.2946 65.0195C85.701 64.7848 86.1167 64.6275 86.5416 64.5475C86.9666 64.4675 87.3823 64.4541 87.7887 64.5075C88.1998 64.5581 88.5832 64.6675 88.9388 64.8355L87.9411 65.4755C87.7564 65.3848 87.5624 65.3261 87.3592 65.2995C87.1513 65.2701 86.9458 65.2741 86.7426 65.3115C86.5393 65.3488 86.343 65.4221 86.1537 65.5315C85.9135 65.6701 85.7772 65.8101 85.7449 65.9515C85.7172 66.0901 85.7888 66.2088 85.9597 66.3075C86.0798 66.3768 86.2114 66.4181 86.3546 66.4315C86.4978 66.4395 86.671 66.4195 86.8742 66.3715C87.082 66.3208 87.3476 66.2341 87.6709 66.1115C88.2206 65.9008 88.6986 65.7608 89.1051 65.6915C89.5115 65.6168 89.8672 65.6035 90.172 65.6515C90.4769 65.6995 90.754 65.7955 91.0034 65.9395C91.3082 66.1155 91.4953 66.3141 91.5646 66.5355C91.6339 66.7515 91.5854 66.9768 91.4191 67.2115C91.2574 67.4435 90.9826 67.6715 90.5946 67.8955ZM94.0076 65.797L89.9754 63.469L88.5067 64.317L87.6198 63.805L91.6243 61.493L92.5112 62.005L91.0354 62.857L95.0677 65.185L94.0076 65.797ZM98.3338 63.2993L96.4216 62.1953L91.7035 61.4473L92.8674 60.7753L95.9228 61.2993L95.0082 59.5393L96.1722 58.8673L97.4747 61.5873L99.3869 62.6913L98.3338 63.2993ZM101.554 61.4398L96.6354 58.5998L97.6885 57.9918L102.323 60.6678L101.139 60.6558L103.986 59.0118L104.873 59.5238L101.554 61.4398ZM105.411 59.2133L100.492 56.3733L101.545 55.7653L106.464 58.6053L105.411 59.2133ZM107.393 58.0687L102.474 55.2287L103.638 54.5567L108.987 55.3887L105.592 53.4287L106.645 52.8207L111.564 55.6607L110.386 56.3407L105.169 55.5687L108.446 57.4607L107.393 58.0687ZM114.643 54.0112C114.167 54.2859 113.652 54.4579 113.098 54.5272C112.544 54.5966 111.982 54.5739 111.414 54.4592C110.846 54.3392 110.31 54.1339 109.807 53.8432C109.313 53.5579 108.964 53.2499 108.761 52.9192C108.558 52.5886 108.518 52.2592 108.643 51.9312C108.768 51.5979 109.075 51.2899 109.564 51.0072C109.989 50.7619 110.419 50.5966 110.853 50.5112C111.287 50.4206 111.721 50.3992 112.156 50.4472C112.59 50.4952 113.017 50.6006 113.437 50.7632L112.398 51.4192C112.075 51.2912 111.751 51.2299 111.428 51.2352C111.105 51.2352 110.781 51.3286 110.458 51.5152C110.176 51.6779 110.015 51.8566 109.973 52.0512C109.936 52.2432 109.999 52.4392 110.16 52.6392C110.327 52.8366 110.571 53.0286 110.895 53.2152C111.223 53.4046 111.557 53.5472 111.899 53.6432C112.241 53.7339 112.576 53.7672 112.904 53.7432C113.236 53.7166 113.55 53.6179 113.846 53.4472C114.063 53.3219 114.206 53.1886 114.276 53.0472C114.35 52.9032 114.354 52.7619 114.289 52.6232C114.229 52.4819 114.112 52.3499 113.936 52.2272L112.814 52.8752L112.045 52.4312L114.199 51.1872L116.756 52.6632L116.105 53.0392L114.996 52.4632L115.176 52.4232C115.347 52.5752 115.437 52.7446 115.446 52.9312C115.456 53.1179 115.393 53.3059 115.259 53.4952C115.125 53.6792 114.92 53.8512 114.643 54.0112Z"
+            fill="black"
+          />
+          <path
+            d="M114.687 29.3232L105.348 34.7154C104.257 35.3454 103.644 36.1999 103.644 37.0908C103.644 37.9818 104.257 38.8362 105.348 39.4663L114.687 44.8585"
+            stroke="black"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+          <path
+            d="M60.7891 32.7432L64.204 34.7148C65.8408 35.6598 68.0608 36.1907 70.3756 36.1907C72.6903 36.1907 74.9103 35.6598 76.5471 34.7148C77.6383 34.0848 79.1183 33.7308 80.6615 33.7308C82.2047 33.7308 83.6847 34.0848 84.7759 34.7148C87.0388 36.0213 90.7417 36.0213 93.0046 34.7148C94.0958 34.0848 94.7088 33.2303 94.7088 32.3393C94.7088 31.4484 94.0958 30.5939 93.0046 29.9639C90.7417 28.6574 90.7417 26.5195 93.0046 25.213L99.5259 21.448"
+            stroke="black"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+          <path
+            d="M107.303 50.0961L99.1774 45.4047C98.0862 44.7747 96.6062 44.4207 95.063 44.4207C93.5198 44.4207 92.0398 44.7747 90.9486 45.4047C89.8574 46.0347 88.3774 46.3886 86.8342 46.3886C85.291 46.3886 83.8111 46.0347 82.7199 45.4047L80.6627 44.2169C79.5715 43.5869 78.0915 43.233 76.5483 43.233C75.0051 43.233 73.5251 43.5869 72.4339 44.2169L66.3652 47.7207"
+            stroke="black"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+          <path
+            d="M109.422 48.9904C120.806 42.4433 120.841 31.8081 109.501 25.2361C98.1608 18.664 79.7402 18.6438 68.3571 25.1909C56.9739 31.7379 56.9389 42.3731 68.2788 48.9452C79.6186 55.5172 98.0393 55.5375 109.422 48.9904Z"
+            stroke="black"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+        </g>
+        <rect
+          width="169"
+          height="40"
+          transform="matrix(0.866025 0.5 -0.866025 0.5 168.413 271.044)"
+          fill="white"
+          stroke="black"
+        />
+        <rect
+          width="169"
+          height="186"
+          transform="matrix(0.866025 0.5 0 1 133.771 291.044)"
+          fill="url(#paint14_linear_4_2)"
+          stroke="url(#paint15_linear_4_2)"
+        />
+        <rect
+          width="40"
+          height="186"
+          transform="matrix(0.866025 -0.5 0 1 280.13 375.544)"
+          fill="url(#paint16_linear_4_2)"
+          stroke="url(#paint17_linear_4_2)"
+        />
+        <path d="M350.673 376.364L315.84 356.252L281.199 376.252" stroke="black" />
+        <path
+          d="M281.738 376.235L315.827 356.327L350.326 376.546L346.143 376.546C346.143 376.053 345.595 375.796 345.51 375.729C336.656 370.565 318.76 360.09 317.619 359.408C316.822 358.932 315.753 358.808 314.733 359.121L285.886 376.02L281.738 376.235Z"
+          fill="white"
+        />
+        <path
+          className="buildingCubeFace"
+          d="M168.581 263.222L198.892 280.722L168.581 305.061L138.27 287.545L168.581 263.222Z"
+          fill="white"
+          stroke="black"
+        />
+        <path
+          className="buildingCubeFace"
+          d="M168.581 265.639L195.167 280.988L168.314 302.49L141.729 287.126L168.581 265.639Z"
+          fill="white"
+          stroke="black"
+        />
+        <rect
+          className="buildingCubeFace"
+          width="35"
+          height="3.18346"
+          transform="matrix(0.866025 0.5 0 1 138.27 287.538)"
+          fill="white"
+          stroke="black"
+        />
+        <path
+          className="buildingCubeFace"
+          d="M168.582 305.049L198.892 280.722V290.722L168.581 308.222L168.582 305.049Z"
+          fill="white"
+          stroke="black"
+        />
+  </>
+);
+
+export default SceneAnimatedLayers;
