+++ components/landing/svg/hero/scene-cargo-and-rocket.tsx
@@ -1,0 +1,274 @@
+import { EMBERS } from "./constants";
+
+const SceneCargoAndRocket = () => (
+  <>
+        <rect
+          width="100"
+          height="93.2218"
+          transform="matrix(0.866025 0.5 0 1 0.865234 93.2266)"
+          fill="url(#paint0_linear_4_2)"
+          stroke="url(#paint1_linear_4_2)"
+        />
+        <rect
+          width="100"
+          height="90.222"
+          transform="matrix(0.866025 -0.5 0 1 87.4678 143.227)"
+          fill="url(#paint2_linear_4_2)"
+          stroke="url(#paint3_linear_4_2)"
+        />
+        <path
+          className="Beam"
+          d="M342.225 403.808C363.292 447.008 413.643 521.361 527.365 463.818C538.857 455.945 562.861 436.07 566.947 427.218"
+          stroke="#1346E7"
+          strokeWidth="4"
+          style={{ opacity: 0 }}
+        />
+        <path
+          className="Beam"
+          d="M339.032 421.685C360.1 464.886 424.154 539.153 560.349 480.206"
+          stroke="#1346E7"
+          strokeWidth="4"
+          style={{ opacity: 0 }}
+        />
+        <rect
+          width="210"
+          height="210"
+          transform="matrix(0.866025 0.5 -0.866025 0.5 352.899 164.696)"
+          fill="white"
+          stroke="black"
+        />
+        <rect
+          width="210"
+          height="167.4"
+          transform="matrix(0.866025 0.5 0 1 171.034 269.696)"
+          fill="url(#paint4_linear_4_2)"
+          stroke="url(#paint5_linear_4_2)"
+        />
+        <rect
+          width="210"
+          height="168.177"
+          transform="matrix(0.866025 -0.5 0 1 352.9 374.696)"
+          fill="url(#paint6_linear_4_2)"
+          stroke="url(#paint7_linear_4_2)"
+        />
+        <rect
+          width="170"
+          height="170"
+          transform="matrix(0.866025 0.5 -0.866025 0.5 352.9 175.68)"
+          fill="#EDEDED"
+          stroke="black"
+        />
+        <rect
+          width="170"
+          height="15"
+          transform="matrix(0.866025 0.5 0 1 205.676 260.68)"
+          fill="white"
+          stroke="black"
+        />
+        <rect
+          width="170"
+          height="15"
+          transform="matrix(0.866025 -0.5 0 1 352.9 345.68)"
+          fill="white"
+          stroke="black"
+        />
+        <circle
+          cx="47"
+          cy="47"
+          r="47"
+          transform="matrix(0.866025 0.5 -0.866025 0.5 682.39 419.754)"
+          fill="white"
+          stroke="black"
+        />
+        <path
+          d="M624.827 466.754C624.827 475.931 631.27 484.24 641.687 490.254V569.254C631.27 563.24 624.827 554.931 624.827 545.754V466.754Z"
+          fill="url(#paint8_linear_4_2)"
+          stroke="url(#paint9_linear_4_2)"
+        />
+        <path
+          d="M641.687 490.254C664.166 503.233 700.613 503.233 723.093 490.254V569.254C700.613 582.233 664.166 582.233 641.687 569.254V490.254Z"
+          fill="url(#paint10_linear_4_2)"
+          stroke="url(#paint11_linear_4_2)"
+        />
+        <path
+          d="M739.953 466.754C739.953 475.931 733.51 484.24 723.093 490.254V569.254C733.51 563.24 739.953 554.931 739.953 545.754V466.754Z"
+          fill="url(#paint12_linear_4_2)"
+          stroke="url(#paint13_linear_4_2)"
+        />
+        <path
+          d="M653.563 411.58L617.798 433.115L611.207 429.143L605.071 425.354L597.726 421.019L592.715 418L586.903 414.497L580.961 410.917L574.658 407.118L569.391 403.944L562.779 399.96L556.609 396.241L550.649 392.649L543.778 388.509L537.751 384.877L532.098 381.47L526.165 377.895L520.137 374.263L514.244 370.711L507.896 366.886L501.855 363.245L495.239 359.258L489.627 355.876L483.895 352.422L478.096 348.927L471.465 344.931L465.33 341.234L459.746 337.869L453.665 334.204L446.901 330.128L441.142 326.658L435.23 323.095L428.975 319.326L423.062 315.762L417.148 312.198L411.018 308.505L404.791 304.752L399.217 301.393L435.277 280.573L440.863 283.926L447.103 287.671L453.246 291.357L459.172 294.914L465.097 298.47L471.365 302.232L477.29 305.788L483.061 309.251L489.839 313.319L495.933 316.976L501.529 320.335L506.942 323.583L514.321 328.012L520.132 331.5L525.876 334.947L531.5 338.323L538.131 342.302L544.184 345.935L550.545 349.753L556.451 353.297L562.491 356.922L568.437 360.49L574.101 363.89L580.141 367.515L587.026 371.647L592.999 375.232L599.182 378.942L605.807 382.919L611.086 386.087L617.402 389.877L623.356 393.451L629.181 396.946L634.202 399.96L641.485 404.331L647.711 408.068L653.563 411.58Z"
+          fill="white"
+        />
+        <path
+          d="M653.563 411.58L617.798 433.115L611.207 429.143L605.071 425.354L597.726 421.019L592.715 418L586.903 414.497L580.961 410.917L574.658 407.118L569.391 403.944L562.779 399.96L556.609 396.241L550.649 392.649L543.778 388.509L537.751 384.877L532.098 381.47L526.165 377.895L520.137 374.263L514.244 370.711L507.896 366.886L501.855 363.245L495.239 359.258L489.627 355.876L483.895 352.422L478.096 348.927L471.465 344.931L465.33 341.234L459.746 337.869L453.665 334.204L446.901 330.128L441.142 326.658L435.23 323.095L428.975 319.326L423.062 315.762L417.148 312.198L411.018 308.505L404.791 304.752L399.217 301.393L435.277 280.573L440.863 283.926L447.103 287.671L453.246 291.357L459.172 294.914L465.097 298.47L471.365 302.232L477.29 305.788L483.061 309.251L489.839 313.319L495.933 316.976L501.529 320.335L506.942 323.583L514.321 328.012L520.132 331.5L525.876 334.947L531.5 338.323L538.131 342.302L544.184 345.935L550.545 349.753L556.451 353.297L562.491 356.922L568.437 360.49L574.101 363.89L580.141 367.515L587.026 371.647L592.999 375.232L599.182 378.942L605.807 382.919L611.086 386.087L617.402 389.877L623.356 393.451L629.181 396.946L634.202 399.96L641.485 404.331L647.711 408.068L653.563 411.58Z"
+          fill="white"
+          fillOpacity="0.2"
+        />
+        <path
+          d="M653.563 411.58L617.798 433.115L611.207 429.143L605.071 425.354L597.726 421.019L592.715 418L586.903 414.497L580.961 410.917L574.658 407.118L569.391 403.944L562.779 399.96L556.609 396.241L550.649 392.649L543.778 388.509L537.751 384.877L532.098 381.47L526.165 377.895L520.137 374.263L514.244 370.711L507.896 366.886L501.855 363.245L495.239 359.258L489.627 355.876L483.895 352.422L478.096 348.927L471.465 344.931L465.33 341.234L459.746 337.869L453.665 334.204L446.901 330.128L441.142 326.658L435.23 323.095L428.975 319.326L423.062 315.762L417.148 312.198L411.018 308.505L404.791 304.752L399.217 301.393L435.277 280.573L440.863 283.926L447.103 287.671L453.246 291.357L459.172 294.914L465.097 298.47L471.365 302.232L477.29 305.788L483.061 309.251L489.839 313.319L495.933 316.976L501.529 320.335L506.942 323.583L514.321 328.012L520.132 331.5L525.876 334.947L531.5 338.323L538.131 342.302L544.184 345.935L550.545 349.753L556.451 353.297L562.491 356.922L568.437 360.49L574.101 363.89L580.141 367.515L587.026 371.647L592.999 375.232L599.182 378.942L605.807 382.919L611.086 386.087L617.402 389.877L623.356 393.451L629.181 396.946L634.202 399.96L641.485 404.331L647.711 408.068L653.563 411.58Z"
+          fill="white"
+          fillOpacity="0.2"
+        />
+        <path
+          d="M653.563 411.58L617.798 433.115L611.207 429.143L605.071 425.354L597.726 421.019L592.715 418L586.903 414.497L580.961 410.917L574.658 407.118L569.391 403.944L562.779 399.96L556.609 396.241L550.649 392.649L543.778 388.509L537.751 384.877L532.098 381.47L526.165 377.895L520.137 374.263L514.244 370.711L507.896 366.886L501.855 363.245L495.239 359.258L489.627 355.876L483.895 352.422L478.096 348.927L471.465 344.931L465.33 341.234L459.746 337.869L453.665 334.204L446.901 330.128L441.142 326.658L435.23 323.095L428.975 319.326L423.062 315.762L417.148 312.198L411.018 308.505L404.791 304.752L399.217 301.393L435.277 280.573L440.863 283.926L447.103 287.671L453.246 291.357L459.172 294.914L465.097 298.47L471.365 302.232L477.29 305.788L483.061 309.251L489.839 313.319L495.933 316.976L501.529 320.335L506.942 323.583L514.321 328.012L520.132 331.5L525.876 334.947L531.5 338.323L538.131 342.302L544.184 345.935L550.545 349.753L556.451 353.297L562.491 356.922L568.437 360.49L574.101 363.89L580.141 367.515L587.026 371.647L592.999 375.232L599.182 378.942L605.807 382.919L611.086 386.087L617.402 389.877L623.356 393.451L629.181 396.946L634.202 399.96L641.485 404.331L647.711 408.068L653.563 411.58Z"
+          fill="white"
+          fillOpacity="0.2"
+        />
+        <path
+          d="M611.207 429.143L617.798 433.115L653.563 411.58L647.711 408.068M611.207 429.143L647.711 408.068M611.207 429.143L605.071 425.354M647.711 408.068L641.485 404.331M605.071 425.354L597.726 421.019M605.071 425.354L641.485 404.331M597.726 421.019L634.202 399.96M597.726 421.019L592.715 418M634.202 399.96L629.181 396.946M634.202 399.96L641.485 404.331M592.715 418L629.181 396.946M592.715 418L586.903 414.497M629.181 396.946L623.356 393.451M586.903 414.497L623.356 393.451M586.903 414.497L580.961 410.917M623.356 393.451L617.402 389.877M580.961 410.917L617.402 389.877M580.961 410.917L574.658 407.118M617.402 389.877L611.086 386.087M574.658 407.118L611.086 386.087M574.658 407.118L569.391 403.944M611.086 386.087L605.807 382.919M569.391 403.944L605.807 382.919M569.391 403.944L562.779 399.96M605.807 382.919L599.182 378.942M562.779 399.96L599.182 378.942M562.779 399.96L556.609 396.241M599.182 378.942L592.999 375.232M556.609 396.241L592.999 375.232M556.609 396.241L550.649 392.649M592.999 375.232L587.026 371.647M550.649 392.649L587.026 371.647M550.649 392.649L543.778 388.509M587.026 371.647L580.141 367.515M543.778 388.509L580.141 367.515M543.778 388.509L537.751 384.877M580.141 367.515L574.101 363.89M537.751 384.877L574.101 363.89M537.751 384.877L532.098 381.47M574.101 363.89L568.437 360.49M532.098 381.47L568.437 360.49M532.098 381.47L526.165 377.895M568.437 360.49L562.491 356.922M526.165 377.895L562.491 356.922M526.165 377.895L520.137 374.263M562.491 356.922L556.451 353.297M520.137 374.263L556.451 353.297M520.137 374.263L514.244 370.711M556.451 353.297L550.545 349.753M514.244 370.711L550.545 349.753M514.244 370.711L507.896 366.886M550.545 349.753L544.184 345.935M507.896 366.886L544.184 345.935M507.896 366.886L501.855 363.245M544.184 345.935L538.131 342.302M501.855 363.245L538.131 342.302M501.855 363.245L495.239 359.258M538.131 342.302L531.5 338.323M495.239 359.258L531.5 338.323M495.239 359.258L489.627 355.876M531.5 338.323L525.876 334.947M489.627 355.876L525.876 334.947M489.627 355.876L483.895 352.422M525.876 334.947L520.132 331.5M483.895 352.422L520.132 331.5M483.895 352.422L478.096 348.927M520.132 331.5L514.321 328.012M478.096 348.927L514.321 328.012M478.096 348.927L471.465 344.931M514.321 328.012L506.942 323.583M471.465 344.931L506.942 323.583M471.465 344.931L465.33 341.234M506.942 323.583L501.529 320.335M465.33 341.234L501.529 320.335M465.33 341.234L459.746 337.869M501.529 320.335L495.933 316.976M459.746 337.869L495.933 316.976M459.746 337.869L453.665 334.204M495.933 316.976L489.839 313.319M453.665 334.204L489.839 313.319M453.665 334.204L446.901 330.128M489.839 313.319L483.061 309.251M446.901 330.128L483.061 309.251M446.901 330.128L441.142 326.658M483.061 309.251L477.29 305.788M441.142 326.658L477.29 305.788M441.142 326.658L435.23 323.095M477.29 305.788L471.365 302.232M435.23 323.095L471.365 302.232M435.23 323.095L428.975 319.326M471.365 302.232L465.097 298.47M428.975 319.326L465.097 298.47M428.975 319.326L423.062 315.762M465.097 298.47L459.172 294.914M423.062 315.762L459.172 294.914M423.062 315.762L417.148 312.198M459.172 294.914L453.246 291.357M417.148 312.198L453.246 291.357M417.148 312.198L411.018 308.505M453.246 291.357L447.103 287.671M411.018 308.505L447.103 287.671M411.018 308.505L404.791 304.752M447.103 287.671L440.863 283.926M404.791 304.752L440.863 283.926M404.791 304.752L399.217 301.393L435.277 280.573L440.863 283.926"
+          stroke="black"
+        />
+        <g clipPath="url(#rocketFlameClip)">
+        <g className="rocket">
+          <g className="rocketFlame" aria-hidden="true">
+            <path
+              className="rocketTrail-halo"
+              d="M628.5 476 Q665 600 658.5 820 L708.5 820 Q702 600 738.5 476 Z"
+              fill="url(#rocketHaloGradient)"
+              filter="url(#rocketHaloBlur)"
+              style={{ opacity: 0 }}
+            />
+            <path
+              className="rocketTrail-plume"
+              d="M644.5 478 Q670 600 669.5 740 L697.5 740 Q697 600 722.5 478 Z"
+              fill="url(#rocketPlumeGradient)"
+              style={{ opacity: 0 }}
+            />
+            <path
+              className="rocketTrail-core"
+              data-nozzle="left"
+              d="M650 475 Q652 535 657 590 Q662 535 664 475 Z"
+              fill="url(#rocketCoreGradient)"
+              style={{ opacity: 0 }}
+            />
+            <path
+              className="rocketTrail-core"
+              data-nozzle="center"
+              d="M677 475 Q680 545 684 615 Q688 545 691 475 Z"
+              fill="url(#rocketCoreGradient)"
+              style={{ opacity: 0 }}
+            />
+            <path
+              className="rocketTrail-core"
+              data-nozzle="right"
+              d="M703 475 Q705 535 710 590 Q715 535 717 475 Z"
+              fill="url(#rocketCoreGradient)"
+              style={{ opacity: 0 }}
+            />
+            <g
+              className="rocketEmbers"
+              aria-hidden="true"
+              style={{ willChange: "transform, opacity" }}
+            >
+              {EMBERS.map((e) => (
+                <circle
+                  key={e.id}
+                  className="rocketEmber"
+                  cx={e.cx}
+                  cy={e.cy}
+                  r={e.radius}
+                  fill={e.fill}
+                  opacity={0}
+                />
+              ))}
+            </g>
+          </g>
+          <path
+            d="M661.605 262.632V247.135C663.194 248.879 665.407 250.775 668.104 252.421C668.896 252.904 669.731 253.366 670.603 253.796C677.357 257.126 686.403 258.554 696.098 253.358C697.092 252.826 698.092 252.224 699.098 251.547C701.247 250.101 703.419 248.313 705.596 246.136V260.633V267.631C708.43 269.692 709.618 272.417 710.016 274.63C710.236 275.852 710.214 276.918 710.095 277.629C710.075 277.712 710.054 277.793 710.033 277.874L710.143 299.125L710.359 340.617L710.437 355.614L710.46 360.113L710.486 365.112L710.541 375.61L710.568 380.859L710.595 386.108L723.093 411.603V470.092C721.26 472.591 716.494 476.99 712.095 474.591V480.59C704.856 486.22 697.962 485.874 693.099 483.98C690.708 483.048 688.808 481.743 687.6 480.59C684.4 482.189 681.268 481.256 680.101 480.59C673.303 486.988 663.938 485.922 660.105 484.589C656.773 483.422 650.807 479.79 653.607 474.591C647.208 473.791 644.608 471.258 644.108 470.092V411.603L657.391 385.108V380.859V376.61V372.111V366.112V361.113V355.614V340.617V299.125V275.63C657.331 270.739 659.592 268.001 661.605 266.132V262.632Z"
+            fill="white"
+          />
+          <path
+            d="M661.605 236.138V247.135M661.605 236.138C663.883 237.495 663.804 237.814 667.604 239.091M661.605 236.138C659.689 235.283 655.978 232.962 656.458 230.51C656.8 228.757 656.876 220.068 660.35 204.448M661.605 247.135V262.632M661.605 247.135C663.194 248.879 665.407 250.775 668.104 252.421M661.605 262.632V266.132M661.605 262.632C662.607 264.252 664.823 266.513 668.104 268.427M705.596 260.633V246.136M705.596 260.633V267.631M705.596 260.633C703.893 264.039 701.51 266.439 698.831 268.131M705.596 246.136V235.341M705.596 246.136C703.419 248.313 701.247 250.101 699.098 251.547M705.596 235.341C702.743 237.654 698.098 239.57 696.598 239.57M705.596 235.341C708.52 233.478 709.134 231.344 709.075 230.51C707.929 221.632 706.237 212.957 704.262 204.752M661.605 266.132V274.63C661.936 277.5 664.286 282.229 670.603 284.977M661.605 266.132C659.592 268.001 657.331 270.739 657.391 275.63M710.033 277.874C709.027 281.72 707.323 283.927 699.098 288.627C690.699 293.426 677.602 291.96 672.103 290.627C667.604 288.627 658.375 283.129 657.606 278.129C657.469 277.238 657.4 276.407 657.391 275.63M710.033 277.874C710.054 277.793 710.075 277.712 710.095 277.629C710.214 276.918 710.236 275.852 710.016 274.63M710.033 277.874L710.143 299.125M710.033 277.874L710.016 274.63M705.596 267.631V274.63C705.293 277.059 703.293 280.999 698.098 283.809M705.596 267.631C708.43 269.692 709.618 272.417 710.016 274.63M657.391 275.63V299.125M657.391 385.108C665.961 393.773 688.6 406.104 710.595 386.108M657.391 385.108V380.859M657.391 385.108L644.108 411.603M657.391 385.108L656.106 416.602M710.595 386.108L710.568 380.859M710.595 386.108L723.093 411.603M710.595 386.108L710.382 416.602M657.391 299.125C670.603 311.123 689.786 315.922 710.143 299.125M657.391 299.125V340.617M710.143 299.125L710.359 340.617M657.391 340.617C665.637 348.282 687.774 359.013 710.359 340.617M657.391 340.617V355.614M710.359 340.617L710.437 355.614M657.391 355.614C666.355 362.946 689.514 373.211 710.437 355.614M657.391 355.614V361.113M710.437 355.614L710.46 360.113M657.391 361.113C666.371 368.278 689.556 378.11 710.46 360.113M657.391 361.113V366.112M710.46 360.113L710.486 365.112M657.391 366.112C665.388 373.111 687.203 382.709 710.486 365.112M657.391 366.112V372.111M710.486 365.112L710.541 375.61M657.391 372.111C665.803 378.443 688.174 387.008 710.359 370.611M657.391 372.111V376.61M657.391 376.61C665.924 383.275 688.501 392.407 710.541 375.61M657.391 376.61V380.859M710.541 375.61L710.568 380.859M657.391 380.859C666.942 388.274 690.95 398.656 710.568 380.859M670.603 253.796C670.77 251.076 671.003 245.536 670.603 245.136C670.103 244.636 667.604 244.136 667.604 243.136V239.091M670.603 253.796C669.731 253.366 668.896 252.904 668.104 252.421M670.603 253.796C677.357 257.126 686.403 258.554 696.098 253.358M667.604 239.091C679.204 241.485 684.869 241.644 696.598 239.57M696.598 239.57C696.431 241.403 696.198 244.336 696.598 245.136C697.098 246.136 699.098 244.636 699.098 246.136V251.547M699.098 251.547C698.092 252.224 697.092 252.826 696.098 253.358M668.104 252.421C667.937 254.158 667.704 257.833 668.104 258.633C668.604 259.633 670.603 259.133 670.603 261.133V269.71M670.603 269.71C674.094 271.277 678.466 272.344 683.601 272.131C687.212 271.98 691.676 271.479 695.81 269.71M670.603 269.71C669.712 269.311 668.878 268.878 668.104 268.427M696.098 253.358C695.932 255.283 695.698 259.333 696.098 260.133C696.598 261.133 699.098 260.133 699.098 260.633C699.098 261.033 698.92 265.798 698.831 268.131M698.831 268.131C697.857 268.747 696.843 269.268 695.81 269.71M668.104 268.427C667.937 269.661 667.704 272.43 668.104 273.63C668.604 275.13 670.603 274.63 670.603 275.63V284.977M670.603 284.977C673.821 286.377 678.067 287.263 683.601 287.128C690.04 286.97 694.738 285.626 698.098 283.809M695.81 269.71C695.713 271.158 695.579 274.268 695.81 275.13C696.098 276.207 697.782 275.273 698.098 276.451C698.35 277.394 698.203 281.749 698.098 283.809M644.108 411.603C645.608 413.936 650.107 418.202 656.106 416.602M644.108 411.603V470.092C644.608 471.258 647.208 473.791 653.607 474.591M656.106 416.602V419.774M653.607 474.591V424.601M653.607 474.591C650.807 479.79 656.773 483.422 660.105 484.589M653.607 424.601C654.806 427 658.439 426.6 660.105 426.1M653.607 424.601L656.106 419.774M668.104 396.606L680.101 423.601M668.104 396.606L660.105 426.1M668.104 396.606L656.106 419.774M680.101 423.601C676.435 426.1 667.304 430.1 660.105 426.1M680.101 423.601V480.59M660.105 426.1V484.589M660.105 484.589C663.938 485.922 673.303 486.988 680.101 480.59M680.101 480.59C681.268 481.256 684.4 482.189 687.6 480.59M687.6 423.601L699.597 396.606M687.6 423.601C687.6 424.401 690.266 425.6 691.599 426.1C692.107 426.3 692.607 426.482 693.099 426.647M687.6 423.601V480.59M699.597 396.606L693.099 426.647M699.597 396.606L710.359 419.851M712.095 423.601C710.027 426.146 703.434 430.118 693.099 426.647M712.095 423.601L710.359 419.851M712.095 423.601V474.591M693.099 426.647V483.98M687.6 480.59C688.808 481.743 690.708 483.048 693.099 483.98M693.099 483.98C697.962 485.874 704.856 486.22 712.095 480.59V474.591M710.359 419.851L710.382 416.602M723.093 411.603C721.117 413.936 715.81 418.202 710.382 416.602M723.093 411.603V470.092C721.26 472.591 716.494 476.99 712.095 474.591M660.35 204.448C668.39 209.463 688.429 216.545 704.262 204.752M660.35 204.448C660.977 201.628 661.714 198.583 662.584 195.311M704.262 204.752C703.485 201.527 702.665 198.375 701.817 195.311M662.584 195.311C669.73 199.707 687.579 205.86 701.817 195.311M662.584 195.311C663.699 191.12 665.031 186.559 666.625 181.627M701.817 195.311C700.489 190.512 699.091 185.929 697.687 181.627M666.625 181.627C672.47 184.835 686.864 189.327 697.687 181.627M666.625 181.627C667.396 179.24 668.229 176.767 669.128 174.207M697.687 181.627C696.845 179.049 696.001 176.571 695.167 174.207M669.128 174.207C670.934 169.067 673.009 163.577 675.393 157.737L680.812 145.988C681.351 144.958 682.696 143.516 683.756 145.988C684.432 147.566 689.564 158.321 695.167 174.207M669.128 174.207C674.146 176.202 686.379 178.995 695.167 174.207M662.605 318.122C664.438 319.121 668.104 321.321 668.104 322.121C668.104 323.121 668.604 333.618 668.104 334.118C667.604 334.618 661.605 331.619 661.605 331.119C661.605 330.619 661.105 317.622 662.605 318.122ZM679.102 323.62C681.268 323.454 685.8 323.221 686.6 323.62C687.6 324.12 687.6 336.618 687.1 337.118C686.6 337.618 679.102 338.118 679.102 337.118C679.102 336.118 678.408 323.67 679.102 323.62ZM698.098 322.121C699.764 320.621 703.297 317.722 704.097 318.122C705.096 318.621 705.096 329.619 704.597 330.119C704.097 330.619 699.098 335.118 698.598 334.618C698.098 334.118 697.598 322.563 698.098 322.121Z"
+            stroke="black"
+          />
+          <path
+            d="M705.115 351.13C704.965 351.122 704.752 351.213 704.542 351.519C704.337 351.818 704.183 352.266 704.157 352.792C704.13 353.318 704.239 353.779 704.413 354.097C704.591 354.422 704.794 354.534 704.945 354.542C705.095 354.549 705.308 354.458 705.518 354.152C705.723 353.853 705.877 353.405 705.903 352.879L706.352 352.902L706.345 353.012C706.256 354.113 705.669 354.969 704.99 354.991L704.922 354.991C704.192 354.954 703.648 353.959 703.708 352.769L703.715 352.658C703.807 351.522 704.43 350.645 705.138 350.681L705.206 350.687C705.902 350.779 706.41 351.749 706.352 352.902L705.903 352.879C705.93 352.353 705.821 351.892 705.647 351.574C705.469 351.249 705.266 351.137 705.115 351.13Z"
+            fill="black"
+          />
+          <path
+            d="M705.113 292.142C704.963 292.135 704.75 292.226 704.54 292.532C704.335 292.831 704.181 293.279 704.155 293.804C704.128 294.33 704.237 294.792 704.411 295.11C704.589 295.435 704.792 295.547 704.943 295.554C705.093 295.562 705.306 295.471 705.516 295.165C705.721 294.866 705.875 294.418 705.902 293.892L706.351 293.914L706.343 294.025C706.254 295.125 705.667 295.982 704.988 296.004L704.92 296.003C704.19 295.967 703.646 294.972 703.706 293.782L703.713 293.671C703.805 292.534 704.428 291.658 705.136 291.693L705.204 291.699C705.9 291.792 706.408 292.761 706.351 293.914L705.902 293.892C705.928 293.366 705.819 292.905 705.645 292.587C705.467 292.261 705.264 292.15 705.113 292.142Z"
+            fill="black"
+          />
+          <path
+            d="M702.165 257.155C702.17 256.817 702.081 256.53 701.959 256.341C701.834 256.148 701.707 256.098 701.628 256.097C701.549 256.096 701.422 256.143 701.293 256.332C701.165 256.518 701.069 256.803 701.065 257.14C701.06 257.478 701.149 257.764 701.271 257.954C701.396 258.146 701.522 258.196 701.601 258.198L701.595 258.647C701.043 258.64 700.604 257.963 700.615 257.134C700.626 256.306 701.082 255.641 701.634 255.648C702.187 255.655 702.626 256.332 702.615 257.161L702.608 257.314C702.546 258.069 702.113 258.654 701.595 258.647L701.601 258.198C701.68 258.199 701.808 258.152 701.937 257.963C702.065 257.776 702.161 257.492 702.165 257.155Z"
+            fill="black"
+          />
+          <path
+            d="M702.165 242.157C702.17 241.819 702.081 241.533 701.959 241.343C701.834 241.151 701.707 241.101 701.628 241.1C701.549 241.099 701.422 241.145 701.293 241.335C701.165 241.521 701.069 241.805 701.065 242.143C701.06 242.48 701.149 242.767 701.271 242.956C701.396 243.149 701.522 243.199 701.601 243.2L701.595 243.649C701.043 243.642 700.604 242.965 700.615 242.137C700.626 241.309 701.082 240.643 701.634 240.65C702.187 240.657 702.626 241.335 702.615 242.163L702.608 242.316C702.546 243.072 702.113 243.656 701.595 243.649L701.601 243.2C701.68 243.201 701.808 243.154 701.937 242.965C702.065 242.779 702.161 242.495 702.165 242.157Z"
+            fill="black"
+          />
+          <path
+            d="M702.165 272.15C702.17 271.812 702.081 271.526 701.959 271.336C701.834 271.143 701.707 271.094 701.628 271.092C701.549 271.092 701.422 271.138 701.293 271.327C701.165 271.514 701.069 271.798 701.065 272.135C701.06 272.473 701.149 272.76 701.271 272.949C701.396 273.142 701.522 273.192 701.601 273.193L701.595 273.642C701.043 273.635 700.604 272.958 700.615 272.129C700.626 271.301 701.082 270.636 701.634 270.643C702.187 270.65 702.626 271.327 702.615 272.156L702.608 272.309C702.546 273.064 702.113 273.649 701.595 273.642L701.601 273.193C701.68 273.194 701.808 273.147 701.937 272.958C702.065 272.772 702.161 272.487 702.165 272.15Z"
+            fill="black"
+          />
+          <path
+            d="M661.523 350.43C661.317 350.447 661.092 350.588 660.913 350.899C660.735 351.208 660.631 351.649 660.672 352.145C660.712 352.641 660.886 353.059 661.112 353.335C661.339 353.614 661.584 353.715 661.79 353.698C661.996 353.682 662.222 353.541 662.401 353.23C662.578 352.921 662.682 352.48 662.642 351.984L663.09 351.948C663.181 353.061 662.653 354.02 661.9 354.138L661.827 354.147C661.035 354.211 660.317 353.332 660.223 352.182C660.13 351.032 660.695 350.047 661.487 349.982L661.561 349.979C662.323 349.973 662.999 350.833 663.09 351.948L662.642 351.984C662.601 351.488 662.427 351.07 662.202 350.794C661.974 350.516 661.729 350.413 661.523 350.43Z"
+            fill="black"
+          />
+          <path
+            d="M662.077 292.19C661.871 292.207 661.646 292.348 661.467 292.659C661.289 292.969 661.185 293.409 661.225 293.905C661.266 294.401 661.44 294.819 661.666 295.096C661.893 295.374 662.138 295.476 662.344 295.459C662.55 295.442 662.775 295.302 662.954 294.99C663.132 294.681 663.236 294.24 663.196 293.744L663.644 293.708C663.735 294.822 663.207 295.781 662.454 295.898L662.381 295.907C661.589 295.971 660.871 295.092 660.777 293.942C660.683 292.792 661.249 291.807 662.041 291.742L662.114 291.739C662.876 291.733 663.553 292.594 663.644 293.708L663.196 293.744C663.155 293.248 662.981 292.831 662.755 292.554C662.528 292.276 662.283 292.174 662.077 292.19Z"
+            fill="black"
+          />
+          <path
+            d="M666.175 246.155C666.179 245.817 666.091 245.53 665.968 245.341C665.843 245.148 665.717 245.098 665.638 245.097C665.559 245.096 665.432 245.143 665.302 245.332C665.175 245.518 665.079 245.803 665.074 246.14C665.07 246.478 665.158 246.764 665.281 246.954C665.406 247.146 665.532 247.196 665.611 247.198L665.605 247.647C665.053 247.64 664.614 246.963 664.625 246.134C664.636 245.306 665.092 244.641 665.644 244.648C666.196 244.655 666.635 245.332 666.625 246.161L666.617 246.314C666.556 247.069 666.123 247.654 665.605 247.647L665.611 247.198C665.69 247.199 665.817 247.152 665.947 246.963C666.075 246.776 666.171 246.492 666.175 246.155Z"
+            fill="black"
+          />
+          <path
+            d="M666.175 261.154C666.179 260.816 666.091 260.529 665.968 260.34C665.843 260.147 665.717 260.097 665.638 260.096C665.559 260.095 665.432 260.142 665.302 260.331C665.175 260.518 665.079 260.802 665.074 261.139C665.07 261.477 665.158 261.763 665.281 261.953C665.406 262.145 665.532 262.195 665.611 262.197L665.605 262.646C665.053 262.639 664.614 261.962 664.625 261.133C664.636 260.305 665.092 259.64 665.644 259.647C666.196 259.654 666.635 260.331 666.625 261.16L666.617 261.313C666.556 262.068 666.123 262.653 665.605 262.646L665.611 262.197C665.69 262.198 665.817 262.151 665.947 261.962C666.075 261.775 666.171 261.491 666.175 261.154Z"
+            fill="black"
+          />
+          <path
+            d="M666.175 276.149C666.179 275.812 666.091 275.525 665.968 275.336C665.843 275.143 665.717 275.093 665.638 275.092C665.559 275.091 665.432 275.138 665.302 275.327C665.175 275.513 665.079 275.797 665.074 276.135C665.07 276.473 665.158 276.759 665.281 276.949C665.406 277.141 665.532 277.191 665.611 277.192L665.605 277.642C665.053 277.634 664.614 276.957 664.625 276.129C664.636 275.301 665.092 274.635 665.644 274.642C666.196 274.65 666.635 275.327 666.625 276.155L666.617 276.308C666.556 277.064 666.123 277.648 665.605 277.642L665.611 277.192C665.69 277.193 665.817 277.147 665.947 276.957C666.075 276.771 666.171 276.487 666.175 276.149Z"
+            fill="black"
+          />
+        </g>
+        </g>
+        <rect
+          width="100"
+          height="100"
+          transform="matrix(0.866025 0.5 -0.866025 0.5 87.4688 43.2266)"
+          fill="white"
+          stroke="black"
+        />
+        <path
+          d="M295.338 195.136L258.948 216.146L252.988 212.554L246.117 208.413L240.09 204.781L234.437 201.375L228.504 197.799L222.476 194.167L216.583 190.615L210.235 186.79L204.194 183.149L197.578 179.162L191.966 175.78L186.234 172.326L180.435 168.831L173.804 164.835L167.669 161.139L162.085 157.773L156.004 154.109L149.24 150.032L143.481 146.562L137.569 142.999L131.314 139.23L125.401 135.667L119.487 132.103L113.357 128.409L149.442 107.575L155.585 111.262L161.511 114.818L167.436 118.374L173.704 122.136L179.629 125.692L185.4 129.155L192.178 133.223L198.272 136.881L203.868 140.239L209.281 143.488L216.66 147.917L222.471 151.404L228.215 154.852L233.839 158.227L240.47 162.206L246.523 165.839L252.884 169.657L258.79 173.201L264.83 176.826L270.776 180.395L276.44 183.794L282.48 187.419L289.365 191.551L295.338 195.136Z"
+          fill="white"
+        />
+        <path
+          d="M295.338 195.136L258.948 216.146L252.988 212.554L246.117 208.413L240.09 204.781L234.437 201.375L228.504 197.799L222.476 194.167L216.583 190.615L210.235 186.79L204.194 183.149L197.578 179.162L191.966 175.78L186.234 172.326L180.435 168.831L173.804 164.835L167.669 161.139L162.085 157.773L156.004 154.109L149.24 150.032L143.481 146.562L137.569 142.999L131.314 139.23L125.401 135.667L119.487 132.103L113.357 128.409L149.442 107.575L155.585 111.262L161.511 114.818L167.436 118.374L173.704 122.136L179.629 125.692L185.4 129.155L192.178 133.223L198.272 136.881L203.868 140.239L209.281 143.488L216.66 147.917L222.471 151.404L228.215 154.852L233.839 158.227L240.47 162.206L246.523 165.839L252.884 169.657L258.79 173.201L264.83 176.826L270.776 180.395L276.44 183.794L282.48 187.419L289.365 191.551L295.338 195.136Z"
+          fill="white"
+          fillOpacity="0.2"
+        />
+        <path
+          d="M295.338 195.136L258.948 216.146L252.988 212.554L246.117 208.413L240.09 204.781L234.437 201.375L228.504 197.799L222.476 194.167L216.583 190.615L210.235 186.79L204.194 183.149L197.578 179.162L191.966 175.78L186.234 172.326L180.435 168.831L173.804 164.835L167.669 161.139L162.085 157.773L156.004 154.109L149.24 150.032L143.481 146.562L137.569 142.999L131.314 139.23L125.401 135.667L119.487 132.103L113.357 128.409L149.442 107.575L155.585 111.262L161.511 114.818L167.436 118.374L173.704 122.136L179.629 125.692L185.4 129.155L192.178 133.223L198.272 136.881L203.868 140.239L209.281 143.488L216.66 147.917L222.471 151.404L228.215 154.852L233.839 158.227L240.47 162.206L246.523 165.839L252.884 169.657L258.79 173.201L264.83 176.826L270.776 180.395L276.44 183.794L282.48 187.419L289.365 191.551L295.338 195.136Z"
+          fill="white"
+          fillOpacity="0.2"
+        />
+        <path
+          d="M295.338 195.136L258.948 216.146L252.988 212.554L246.117 208.413L240.09 204.781L234.437 201.375L228.504 197.799L222.476 194.167L216.583 190.615L210.235 186.79L204.194 183.149L197.578 179.162L191.966 175.78L186.234 172.326L180.435 168.831L173.804 164.835L167.669 161.139L162.085 157.773L156.004 154.109L149.24 150.032L143.481 146.562L137.569 142.999L131.314 139.23L125.401 135.667L119.487 132.103L113.357 128.409L149.442 107.575L155.585 111.262L161.511 114.818L167.436 118.374L173.704 122.136L179.629 125.692L185.4 129.155L192.178 133.223L198.272 136.881L203.868 140.239L209.281 143.488L216.66 147.917L222.471 151.404L228.215 154.852L233.839 158.227L240.47 162.206L246.523 165.839L252.884 169.657L258.79 173.201L264.83 176.826L270.776 180.395L276.44 183.794L282.48 187.419L289.365 191.551L295.338 195.136Z"
+          fill="white"
+          fillOpacity="0.2"
+        />
+        <path
+          d="M252.988 212.554L258.948 216.146L295.338 195.136L289.365 191.551M252.988 212.554L289.365 191.551M252.988 212.554L246.117 208.413M289.365 191.551L282.48 187.419M246.117 208.413L282.48 187.419M246.117 208.413L240.09 204.781M282.48 187.419L276.44 183.794M240.09 204.781L276.44 183.794M240.09 204.781L234.437 201.375M276.44 183.794L270.776 180.395M234.437 201.375L270.776 180.395M234.437 201.375L228.504 197.799M270.776 180.395L264.83 176.826M228.504 197.799L264.83 176.826M228.504 197.799L222.476 194.167M264.83 176.826L258.79 173.201M222.476 194.167L258.79 173.201M222.476 194.167L216.583 190.615M258.79 173.201L252.884 169.657M216.583 190.615L252.884 169.657M216.583 190.615L210.235 186.79M252.884 169.657L246.523 165.839M210.235 186.79L246.523 165.839M210.235 186.79L204.194 183.149M246.523 165.839L240.47 162.206M204.194 183.149L240.47 162.206M204.194 183.149L197.578 179.162M240.47 162.206L233.839 158.227M197.578 179.162L233.839 158.227M197.578 179.162L191.966 175.78M233.839 158.227L228.215 154.852M191.966 175.78L228.215 154.852M191.966 175.78L186.234 172.326M228.215 154.852L222.471 151.404M186.234 172.326L222.471 151.404M186.234 172.326L180.435 168.831M222.471 151.404L216.66 147.917M180.435 168.831L216.66 147.917M180.435 168.831L173.804 164.835M216.66 147.917L209.281 143.488M173.804 164.835L209.281 143.488M173.804 164.835L167.669 161.139M209.281 143.488L203.868 140.239M167.669 161.139L203.868 140.239M167.669 161.139L162.085 157.773M203.868 140.239L198.272 136.881M162.085 157.773L198.272 136.881M162.085 157.773L156.004 154.109M198.272 136.881L192.178 133.223M156.004 154.109L192.178 133.223M156.004 154.109L149.24 150.032M192.178 133.223L185.4 129.155M149.24 150.032L185.4 129.155M149.24 150.032L143.481 146.562M185.4 129.155L179.629 125.692M143.481 146.562L179.629 125.692M143.481 146.562L137.569 142.999M179.629 125.692L173.704 122.136M137.569 142.999L173.704 122.136M137.569 142.999L131.314 139.23M173.704 122.136L167.436 118.374M131.314 139.23L167.436 118.374M131.314 139.23L125.401 135.667M167.436 118.374L161.511 114.818M125.401 135.667L161.511 114.818M125.401 135.667L119.487 132.103M161.511 114.818L155.585 111.262M119.487 132.103L155.585 111.262M119.487 132.103L113.357 128.409L149.442 107.575L155.585 111.262"
+          stroke="black"
+        />
+        <rect
+          width="73.1803"
+          height="73.1803"
+          transform="matrix(0.866025 0.5 -0.866025 0.5 87.4688 56.3809)"
+          fill="white"
+          stroke="black"
+        />
+        <path
+          d="M458.491 268.27V219.27L429.574 235.825V257.667L436.205 253.469V280.799L458.491 268.27Z"
+          fill="white"
+        />
+        <path
+          d="M429.456 257.365L436.205 253.469V280.798L458.491 268.269V219.269L429.456 235.659"
+          stroke="black"
+        />
+  </>
+);
+
+export default SceneCargoAndRocket;
