+++ components/landing/svg/hero/scene-conveyor.tsx
@@ -1,0 +1,129 @@
+const SceneConveyor = () => (
+  <>
+        <g className="path1svg1" style={{ opacity: 0 }}>
+          <rect
+            width="30"
+            height="30"
+            transform="matrix(0.866025 0.5 -0.866025 0.5 148.089 112.242)"
+            fill="white"
+            stroke="black"
+          />
+          <path
+            d="M160.368 123.726L156.141 126.166C155.647 126.452 155.369 126.838 155.369 127.242C155.369 127.645 155.647 128.032 156.141 128.317L160.368 130.758"
+            stroke="black"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+          <path
+            d="M135.971 125.274L137.517 126.166C138.257 126.594 139.262 126.834 140.31 126.834C141.358 126.834 142.363 126.594 143.104 126.166C143.598 125.881 144.268 125.721 144.966 125.721C145.665 125.721 146.335 125.881 146.829 126.166C147.853 126.758 149.529 126.758 150.554 126.166C151.047 125.881 151.325 125.494 151.325 125.091C151.325 124.688 151.047 124.301 150.554 124.016C149.529 123.424 149.529 122.457 150.554 121.865L153.506 120.161"
+            stroke="black"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+          <path
+            d="M157.026 133.129L153.348 131.005C152.854 130.72 152.184 130.56 151.486 130.56C150.787 130.56 150.117 130.72 149.623 131.005C149.129 131.29 148.459 131.451 147.761 131.451C147.062 131.451 146.392 131.29 145.898 131.005L144.967 130.468C144.473 130.182 143.803 130.022 143.105 130.022C142.406 130.022 141.736 130.182 141.242 130.468L138.495 132.054"
+            stroke="black"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+          <path
+            d="M157.986 132.628C163.139 129.665 163.155 124.85 158.022 121.876C152.888 118.901 144.55 118.891 139.397 121.855C134.245 124.819 134.229 129.633 139.362 132.608C144.495 135.583 152.833 135.592 157.986 132.628Z"
+            stroke="black"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+        </g>
+
+        <line className="path1" x1="132.996" y1="117.184" x2="315.978" y2="225.675" stroke="#" />
+        <g className="path2svg2" style={{ opacity: 0 }}>
+          <rect
+            width="30"
+            height="30"
+            transform="matrix(0.866025 0.5 -0.866025 0.5 394.783 247.538)"
+            fill="white"
+            stroke="black"
+          />
+          <rect
+            width="30"
+            height="10"
+            transform="matrix(0.866025 0.5 0 1 368.802 262.538)"
+            fill="white"
+            stroke="black"
+          />
+          <rect
+            width="30"
+            height="10"
+            transform="matrix(0.866025 -0.5 0 1 394.783 277.538)"
+            fill="white"
+            stroke="black"
+          />
+          <path
+            d="M407.061 259.022L402.833 261.463C402.339 261.748 402.062 262.135 402.062 262.538C402.062 262.941 402.339 263.328 402.833 263.613L407.061 266.054"
+            stroke="black"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+          <path
+            d="M382.663 260.57L384.209 261.463C384.95 261.891 385.955 262.131 387.003 262.131C388.05 262.131 389.055 261.891 389.796 261.463C390.29 261.178 390.96 261.017 391.659 261.017C392.357 261.017 393.027 261.178 393.521 261.463C394.545 262.054 396.222 262.054 397.246 261.463C397.74 261.178 398.017 260.791 398.017 260.388C398.017 259.984 397.74 259.597 397.246 259.312C396.222 258.721 396.222 257.753 397.246 257.162L400.198 255.457"
+            stroke="black"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+          <path
+            d="M403.719 268.425L400.04 266.302C399.546 266.017 398.877 265.856 398.178 265.856C397.479 265.856 396.809 266.017 396.316 266.302C395.822 266.587 395.152 266.747 394.453 266.747C393.755 266.747 393.085 266.587 392.591 266.302L391.659 265.764C391.166 265.479 390.496 265.319 389.797 265.319C389.098 265.319 388.429 265.479 387.935 265.764L385.188 267.35"
+            stroke="black"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+
+          <path
+            d="M404.678 267.925C409.831 264.961 409.847 260.147 404.714 257.172C399.581 254.197 391.242 254.188 386.09 257.151C380.937 260.115 380.921 264.929 386.054 267.904C391.187 270.879 399.526 270.888 404.678 267.925Z"
+            stroke="black"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+        </g>
+        <path
+          d="M359.828 276.225V325.225L398.542 302.428V276.225L429.737 257.379V235.537L359.828 276.225Z"
+          fill="white"
+        />
+        <path
+          d="M429.737 257.313L398.542 276.224V302.428L359.828 325.224V276.224L429.737 235.491"
+          stroke="black"
+        />
+        <circle
+          cx="2.6373"
+          cy="2.6373"
+          r="2.1373"
+          transform="matrix(0.866025 -0.5 0.866025 0.5 82.9004 134.985)"
+          fill="white"
+          stroke="black"
+        />
+        <circle
+          cx="2.6373"
+          cy="2.6373"
+          r="2.1373"
+          transform="matrix(0.866025 -0.5 0.866025 0.5 8.81152 93.5547)"
+          fill="white"
+          stroke="black"
+        />
+        <circle
+          cx="2.6373"
+          cy="2.6373"
+          r="2.1373"
+          transform="matrix(0.866025 -0.5 0.866025 0.5 82.9004 50.5586)"
+          fill="white"
+          stroke="black"
+        />
+        <circle
+          cx="2.6373"
+          cy="2.6373"
+          r="2.1373"
+          transform="matrix(0.866025 -0.5 0.866025 0.5 158.812 93.5547)"
+          fill="white"
+          stroke="black"
+        />
+  </>
+);
+
+export default SceneConveyor;
