+++ components/landing/svg/hero/scene-cubes.tsx
@@ -1,0 +1,189 @@
+type SceneCubesProps = {
+  labelText: string;
+  launchButtonState: "launch" | "relaunch";
+};
+
+const SceneCubes = ({ labelText, launchButtonState }: SceneCubesProps) => (
+  <>
+        <g className="buildingVectorLabel" style={{ opacity: 0 }}>
+        <path
+          d="M154.8 278.19L158.157 275.568L159.482 276.333C159.925 276.589 160.187 276.838 160.267 277.082C160.347 277.325 160.23 277.57 159.915 277.816C159.773 277.927 159.615 278.005 159.442 278.051C159.269 278.096 159.087 278.107 158.896 278.082C158.711 278.057 158.531 277.996 158.355 277.898L158.368 277.88C158.677 278.067 158.843 278.27 158.865 278.488C158.893 278.706 158.786 278.909 158.543 279.099C158.238 279.338 157.895 279.453 157.515 279.443C157.136 279.434 156.738 279.309 156.322 279.069L154.8 278.19ZM155.82 278.156L156.765 278.702C156.997 278.836 157.22 278.907 157.433 278.914C157.653 278.921 157.847 278.86 158.014 278.729C158.178 278.601 158.24 278.463 158.199 278.316C158.163 278.17 158.028 278.031 157.796 277.897L156.85 277.351L155.82 278.156ZM157.285 277.011L158.039 277.446C158.285 277.588 158.518 277.665 158.738 277.676C158.958 277.688 159.146 277.633 159.3 277.512C159.464 277.384 159.519 277.252 159.467 277.114C159.42 276.976 159.269 276.833 159.013 276.685L158.259 276.25L157.285 277.011ZM159.658 281.098C159.367 280.93 159.163 280.743 159.045 280.539C158.934 280.333 158.915 280.123 158.986 279.907C159.061 279.693 159.228 279.485 159.486 279.284L161.661 277.584L162.222 277.908L160.047 279.608C159.792 279.807 159.671 280.005 159.685 280.201C159.702 280.395 159.849 280.572 160.126 280.732C160.403 280.892 160.686 280.963 160.973 280.945C161.265 280.925 161.538 280.815 161.793 280.616L163.968 278.916L164.529 279.24L162.354 280.94C162.096 281.141 161.813 281.282 161.506 281.362C161.203 281.444 160.892 281.463 160.573 281.421C160.257 281.375 159.952 281.268 159.658 281.098ZM162.061 282.383L165.418 279.76L165.979 280.084L162.622 282.707L162.061 282.383ZM163.518 283.224L166.874 280.601L167.436 280.925L164.268 283.4L164.251 283.011L166.345 284.22L165.877 284.586L163.518 283.224ZM166.319 284.841L169.676 282.218L170.86 282.902C171.439 283.236 171.737 283.607 171.756 284.017C171.778 284.423 171.519 284.837 170.98 285.258C170.444 285.677 169.877 285.91 169.279 285.959C168.684 286.009 168.103 285.871 167.534 285.543L166.319 284.841ZM167.348 284.799L167.971 285.159C168.38 285.395 168.784 285.493 169.184 285.454C169.59 285.415 169.995 285.237 170.398 284.922C170.808 284.602 171.01 284.305 171.004 284.032C171.005 283.758 170.801 283.504 170.392 283.268L169.769 282.908L167.348 284.799ZM169.906 286.912L173.263 284.289L173.824 284.613L170.467 287.236L169.906 286.912ZM171.271 287.7L174.628 285.077L175.34 285.488L174.307 288.611L177.044 286.472L177.606 286.796L174.249 289.419L173.511 288.993L174.508 285.933L171.832 288.024L171.271 287.7ZM176.441 290.788C176.091 290.586 175.865 290.355 175.764 290.095C175.666 289.833 175.682 289.561 175.812 289.281C175.945 288.998 176.182 288.724 176.522 288.458C176.86 288.194 177.226 287.997 177.623 287.866C178.023 287.737 178.425 287.687 178.829 287.715C179.236 287.74 179.617 287.855 179.97 288.059C180.271 288.233 180.477 288.418 180.587 288.614C180.701 288.813 180.733 289.013 180.684 289.216C180.639 289.421 180.531 289.623 180.362 289.821L179.747 289.504C179.91 289.32 179.982 289.136 179.963 288.954C179.946 288.769 179.793 288.592 179.502 288.424C179.245 288.276 178.978 288.201 178.698 288.198C178.419 288.195 178.141 288.247 177.866 288.354C177.595 288.462 177.341 288.609 177.104 288.794C176.858 288.986 176.677 289.179 176.561 289.373C176.449 289.57 176.417 289.759 176.466 289.942C176.522 290.124 176.679 290.289 176.935 290.437C177.136 290.553 177.348 290.622 177.573 290.645C177.797 290.667 178.017 290.648 178.232 290.588C178.451 290.53 178.645 290.439 178.816 290.315L177.829 289.745L178.259 289.409L179.782 290.288L178.07 291.625L177.686 291.403L178.406 290.791L178.422 290.878C178.232 290.974 178.02 291.034 177.786 291.057C177.552 291.08 177.317 291.068 177.081 291.022C176.848 290.978 176.635 290.9 176.441 290.788Z"
+          fill="black"
+        />
+        </g>
+        <text
+          className="buildingLiveLabel"
+          x="167"
+          y="285"
+          textAnchor="middle"
+          fontSize="7"
+          fontWeight="700"
+          letterSpacing="0.4"
+          fill="black"
+          pointerEvents="none"
+          transform="matrix(0.866 0.5 -0.781 0.625 244.96 23.38)"
+          aria-label={`Launch status: ${labelText}`}
+          style={{ opacity: 1 }}
+        >
+          {labelText}
+        </text>
+        <path
+          d="M202.02 282.639L310.273 345.139L279.962 369.422L171.709 306.99L202.02 282.639Z"
+          fill="white"
+          stroke="black"
+        />
+        <path
+          d="M202.269 284.739L306.664 345.011L279.524 367.222L175.132 307.013L202.269 284.739Z"
+          fill="white"
+          stroke="black"
+        />
+        <rect
+          width="125"
+          height="3.17605"
+          transform="matrix(0.866025 0.5 0 1 171.708 306.962)"
+          fill="white"
+          stroke="black"
+        />
+        <path
+          d="M279.962 369.47L310.273 345.139V355.139L279.962 372.639V369.47Z"
+          fill="white"
+          stroke="black"
+        />
+        <g className="LaunchButton">
+          <path
+            d="M315.211 347.992C315.689 347.716 316.465 347.716 316.943 347.992L345.522 364.492C346 364.768 346 365.216 345.522 365.492L316.943 381.992C316.465 382.268 315.689 382.268 315.211 381.992L286.632 365.492C286.154 365.216 286.154 364.768 286.632 364.492L315.211 347.992Z"
+            fill="#84A2FF"
+            stroke="black"
+          />
+          <rect
+            width="33"
+            height="20"
+            transform="matrix(0.866025 0.5 0 1 286.632 365.492)"
+            fill="#84A2FF"
+            stroke="black"
+          />
+          <rect
+            width="33"
+            height="20"
+            transform="matrix(0.866025 -0.5 0 1 316.942 381.992)"
+            fill="#84A2FF"
+            stroke="black"
+          />
+          <path
+            d="M286.273 364.992C286.273 365.187 286.411 365.364 286.632 365.492V385.492C286.411 385.364 286.273 385.187 286.273 384.992V364.992Z"
+            fill="#84A2FF"
+            stroke="black"
+          />
+          <path
+            d="M315.211 381.992C315.689 382.268 316.465 382.268 316.943 381.992V401.992C316.465 402.268 315.689 402.268 315.211 401.992V381.992Z"
+            fill="#84A2FF"
+            stroke="black"
+          />
+          <path
+            d="M345.881 364.992C345.881 365.187 345.744 365.364 345.522 365.492V385.492C345.744 385.364 345.881 385.187 345.881 384.992V364.992Z"
+            fill="#84A2FF"
+            stroke="black"
+          />
+          <text
+            className="launchLiveGlyph launchLiveGlyph_launch"
+            x="314"
+            y="367"
+            textAnchor="middle"
+            fontSize={5.5}
+            fontWeight="700"
+            letterSpacing={0.4}
+            fill="black"
+            pointerEvents="none"
+            transform="matrix(0.866 0.5 -0.866 0.5 360.17 25.5)"
+            style={{ opacity: 1 }}
+            aria-label="Launch"
+            aria-hidden={launchButtonState === "relaunch"}
+          >
+            LAUNCH
+          </text>
+          <text
+            className="launchLiveGlyph launchLiveGlyph_relaunch"
+            x="314"
+            y="367"
+            textAnchor="middle"
+            fontSize={5.3}
+            fontWeight="700"
+            letterSpacing={0}
+            fill="black"
+            pointerEvents="none"
+            transform="matrix(0.866 0.5 -0.866 0.5 360.17 25.5)"
+            style={{ opacity: 0 }}
+            aria-label="Re-launch"
+            aria-hidden={launchButtonState === "launch"}
+          >
+            RE-LAUNCH
+          </text>
+        </g>
+        <g
+          className="relaunchArrow"
+          style={{ opacity: 0 }}
+          aria-hidden="true"
+          pointerEvents="none"
+        >
+          <g
+            transform="translate(306.6 321.4) scale(0.2)"
+            stroke="#0F3D6B"
+            strokeWidth="1"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            vectorEffect="non-scaling-stroke"
+          >
+            <path
+              d="M106.77 78.2998C106.82 80.0298 106.28 81.2098 105.17 81.8498L68.47 102.77L48.47 112.77L65.85 102.86L85.17 91.8498C86.28 91.2098 86.82 90.0298 86.77 88.2998C86.72 86.5698 86.19 84.7798 85.17 82.9098C84.06 80.9898 82.76 79.5798 81.28 78.6698C80.88 78.4198 80.49 78.2298 80.12 78.0998C79.25 77.7898 78.46 77.7898 77.74 78.0898L97.39 68.2598C98.5 67.6198 99.8 67.7598 101.28 68.6698C102.76 69.5798 104.06 70.9898 105.17 72.9098C106.19 74.7798 106.72 76.5698 106.77 78.2998Z"
+              fill="#1565C0"
+              vectorEffect="non-scaling-stroke"
+            />
+            <path
+              d="M86.77 88.2998C86.82 90.0298 86.28 91.2098 85.17 91.8498L65.8499 102.86L48.4699 112.77C47.9199 113.09 47.3099 113.19 46.6599 113.08C46.0199 112.97 45.32 112.7 44.58 112.28C43.84 111.85 43.14 111.31 42.49 110.67C41.85 110.03 41.24 109.23 40.69 108.27L3.98996 44.9798C2.96996 43.2198 2.45996 41.4598 2.45996 39.7098C2.45996 37.9598 2.96996 36.7298 3.98996 36.0398L4.33997 35.8698C5.37997 35.4398 6.57998 35.6198 7.94998 36.4198C9.47998 37.2998 10.8 38.6998 11.91 40.6198L39.02 87.3698V16.0698C39.02 14.2598 39.55 13.0498 40.62 12.4498L40.95 12.2898C41.95 11.8798 43.16 12.0798 44.58 12.8998C46.16 13.8098 47.4799 15.1798 48.5399 17.0198C49.6099 18.8598 50.14 20.6798 50.14 22.4898V93.7898L56.12 90.3798L70.14 82.3898L77.39 78.2598L77.74 78.0898C78.46 77.7898 79.25 77.7898 80.12 78.0998C80.49 78.2298 80.88 78.4198 81.28 78.6698C82.76 79.5798 84.06 80.9898 85.17 82.9098C86.19 84.7798 86.72 86.5698 86.77 88.2998Z"
+              fill="#229EFF"
+              vectorEffect="non-scaling-stroke"
+            />
+            <path
+              d="M70.14 12.4897V82.3897L56.1199 90.3797L50.14 93.7897V22.4897C50.14 20.6797 49.6099 18.8597 48.5399 17.0197C47.4799 15.1797 46.16 13.8097 44.58 12.8997C43.16 12.0797 41.95 11.8797 40.95 12.2897L60.6199 2.44972C61.6799 1.83972 63 1.98972 64.58 2.89972C66.16 3.80972 67.4799 5.17972 68.5399 7.01972C69.6099 8.85972 70.14 10.6797 70.14 12.4897Z"
+              fill="#7AC2FF"
+              vectorEffect="non-scaling-stroke"
+            />
+            <path
+              d="M39.02 42.8797V87.3697L11.91 40.6197C10.8 38.6997 9.47998 37.2997 7.94998 36.4197C6.57998 35.6197 5.37997 35.4397 4.33997 35.8697L23.99 26.0397C25.1 25.4097 26.42 25.5297 27.95 26.4197C29.48 27.2997 30.7999 28.6997 31.9099 30.6197L39.02 42.8797Z"
+              fill="#1565C0"
+              vectorEffect="non-scaling-stroke"
+            />
+          </g>
+        </g>
+        <path
+          d="M346.004 376.59L317.02 393.324C316.542 393.6 315.767 393.6 315.288 393.324L286.229 376.547"
+          stroke="black"
+          strokeLinecap="round"
+        />
+        <path
+          d="M350.326 376.455L316.238 396.363L281.738 376.144L285.922 376.144C285.922 376.637 286.469 376.894 286.554 376.961C295.409 382.125 313.305 392.6 314.445 393.281C315.243 393.758 316.312 393.882 317.331 393.569L346.179 376.67L350.326 376.455Z"
+          fill="white"
+        />
+        <path
+          d="M350.326 376.7L316.285 396.354L281.662 376.364"
+          stroke="black"
+          strokeLinecap="round"
+        />
+        <rect
+          width="40"
+          height="185.45"
+          transform="matrix(0.866025 -0.5 0 1 316.03 396.915)"
+          fill="url(#paint18_linear_4_2)"
+          stroke="url(#paint19_linear_4_2)"
+        />
+        <rect
+          width="40"
+          height="185.612"
+          transform="matrix(0.866025 0.5 0 1 281.389 376.751)"
+          fill="url(#paint20_linear_4_2)"
+          stroke="url(#paint21_linear_4_2)"
+        />
+  </>
+);
+
+export default SceneCubes;
