+++ components/landing/svg/hero/scene-pipes-and-pulse.tsx
@@ -1,0 +1,81 @@
+const ScenePipesAndPulse = () => (
+  <>
+        <path
+          className="Pluse"
+          d="M188.591 296.229L200.793 303.302C201.023 303.436 201.297 303.472 201.554 303.403L213.073 300.304C213.816 300.104 214.497 300.784 214.299 301.528L209.911 317.967C209.713 318.711 210.394 319.391 211.137 319.191L225.11 315.431C225.367 315.362 225.641 315.398 225.872 315.532L245.89 327.136C246.12 327.27 246.394 327.306 246.651 327.237L257.493 324.32C258.237 324.12 258.918 324.8 258.719 325.544L253.634 344.6C253.435 345.343 254.116 346.023 254.86 345.823L265.72 342.901L273.637 340.771C274.272 340.6 274.896 341.079 274.896 341.737V348.572C274.896 348.929 275.086 349.258 275.395 349.437L290.534 358.213"
+          stroke="black"
+        />
+        <foreignObject x="336.094" y="391.609" width="241.075" height="101.726">
+          <div
+            style={{
+              backdropFilter: "blur(2px)",
+              clipPath: "url(#bgblur_0_4_2_clip_path)",
+              height: "100%",
+              width: "100%",
+            }}
+          ></div>
+        </foreignObject>
+        <g filter="url(#filter0_ii_4_2)" data-figma-bg-blur-radius="4">
+          <path
+            className="beamPipeFill"
+            d="M512.232 467.826C412.654 509.509 359.092 441.774 343.505 400.085C343.164 399.173 341.844 399.225 341.565 400.158L340.635 403.278C340.576 403.476 340.581 403.692 340.647 403.888C368.614 486.21 450.308 505.384 514.659 475.657C565.88 451.996 574.485 416.844 572.385 401.932C572.358 401.743 572.277 401.576 572.151 401.432L567.786 396.451C567.124 395.695 565.831 396.289 565.919 397.29C568.935 431.93 531.805 458.598 512.232 467.826Z"
+            fill="url(#beamPipeFlow)"
+            mask="url(#pipeRevealMask)"
+          />
+          <path
+            d="M512.232 467.826C412.654 509.509 359.092 441.774 343.505 400.085C343.164 399.173 341.844 399.225 341.565 400.158L340.635 403.278C340.576 403.476 340.581 403.692 340.647 403.888C368.614 486.21 450.308 505.384 514.659 475.657C565.88 451.996 574.485 416.844 572.385 401.932C572.358 401.743 572.277 401.576 572.151 401.432L567.786 396.451C567.124 395.695 565.831 396.289 565.919 397.29C568.935 431.93 531.805 458.598 512.232 467.826Z"
+            stroke="url(#paint22_linear_4_2)"
+          />
+          <path
+            d="M512.232 467.826C412.654 509.509 359.092 441.774 343.505 400.085C343.164 399.173 341.844 399.225 341.565 400.158L340.635 403.278C340.576 403.476 340.581 403.692 340.647 403.888C368.614 486.21 450.308 505.384 514.659 475.657C565.88 451.996 574.485 416.844 572.385 401.932C572.358 401.743 572.277 401.576 572.151 401.432L567.786 396.451C567.124 395.695 565.831 396.289 565.919 397.29C568.935 431.93 531.805 458.598 512.232 467.826Z"
+            stroke="black"
+          />
+        </g>
+
+        <foreignObject x="331.681" y="408.46" width="282.649" height="102.087">
+          <div
+            style={{
+              backdropFilter: "blur(2px)",
+              clipPath: "url(#bgblur_1_4_2_clip_path)",
+              height: "100%",
+              width: "100%",
+            }}
+          ></div>
+        </foreignObject>
+        <g filter="url(#filter1_ii_4_2)" data-figma-bg-blur-radius="4">
+          <path
+            className="beamPipeFill"
+            d="M538.563 485.041C420.667 526.893 357.473 458.439 339.383 416.795C339.016 415.949 337.797 416.009 337.491 416.879L336.237 420.445C336.156 420.673 336.163 420.926 336.253 421.15C369.249 503.432 465.558 522.594 541.425 492.873C601.785 469.226 611.958 434.104 609.499 419.176C609.465 418.969 609.365 418.789 609.214 418.643L603.634 413.242C602.932 412.562 601.726 413.203 601.834 414.174C605.706 448.976 561.717 475.784 538.563 485.041Z"
+            fill="url(#beamPipeFlow)"
+            mask="url(#pipeRevealMask)"
+          />
+          <path
+            d="M538.563 485.041C420.667 526.893 357.473 458.439 339.383 416.795C339.016 415.949 337.797 416.009 337.491 416.879L336.237 420.445C336.156 420.673 336.163 420.926 336.253 421.15C369.249 503.432 465.558 522.594 541.425 492.873C601.785 469.226 611.958 434.104 609.499 419.176C609.465 418.969 609.365 418.789 609.214 418.643L603.634 413.242C602.932 412.562 601.726 413.203 601.834 414.174C605.706 448.976 561.717 475.784 538.563 485.041Z"
+            stroke="url(#paint23_linear_4_2)"
+          />
+          <path
+            d="M538.563 485.041C420.667 526.893 357.473 458.439 339.383 416.795C339.016 415.949 337.797 416.009 337.491 416.879L336.237 420.445C336.156 420.673 336.163 420.926 336.253 421.15C369.249 503.432 465.558 522.594 541.425 492.873C601.785 469.226 611.958 434.104 609.499 419.176C609.465 418.969 609.365 418.789 609.214 418.643L603.634 413.242C602.932 412.562 601.726 413.203 601.834 414.174C605.706 448.976 561.717 475.784 538.563 485.041Z"
+            stroke="black"
+          />
+        </g>
+        <line
+          x1="132.996"
+          y1="117.184"
+          x2="315.978"
+          y2="225.675"
+          stroke="white"
+          strokeOpacity="0.1"
+        />
+        <line className="path2" x1="413.406" y1="287.442" x2="596.388" y2="395.934" stroke="#" />
+        <path
+          d="M644.528 497.203V539.282L618.127 552.09L588.347 568.528L559.39 552.09V509.973V468.137V426.3V385.629L591.053 323.859V292.593V265.497V238.789V238.401V211.118L618.127 149.163L644.528 164.406V211.693V238.789V260.499L658.794 252.263L668.531 257.885L644.528 271.743V292.982V323.859V337.935L658.794 329.699L668.531 335.321L644.528 349.179V366.437V408.274V450.11V491.947V497.203Z"
+          fill="url(#paint24_linear_4_2)"
+        />
+        <path
+          d="M644.528 164.406L618.127 149.163M644.528 164.406L618.19 226.786M644.528 164.406V211.693M644.528 164.406L591.053 211.118M618.127 226.936L618.19 226.786M618.127 226.936L644.528 211.693M618.127 226.936V149.163M618.127 226.936V254.032M618.127 226.936L644.528 238.789M618.127 226.936L591.053 238.789M618.127 552.09L644.528 539.282M618.127 552.09L588.347 568.528M618.127 552.09V507.189M618.127 552.09L588.347 526.692M618.127 552.09L644.528 491.947M644.528 539.282V497.203V491.947M644.528 539.282L618.127 507.189M618.127 339.31L591.053 323.859M618.127 339.31L588.347 402.198M618.127 339.31L644.528 323.859M618.127 339.31V381.68M618.127 339.31V308.224M618.127 339.31L644.528 292.982M618.127 339.31L559.39 385.629M618.127 339.31L591.053 292.593M618.127 339.31L644.528 366.437M588.347 402.198L559.39 385.629M588.347 402.198L618.127 381.68M588.347 402.198L591.053 323.859M588.347 402.198V443.019M588.347 402.198L618.127 423.517M588.347 402.198L559.39 426.3M588.347 568.528L559.39 552.09M588.347 568.528V526.692M588.347 568.528L618.127 507.189M588.347 568.528L559.39 509.973M644.528 349.179L668.531 335.321L658.794 329.699L644.528 337.935M644.528 349.179V337.935M644.528 349.179V366.437M644.528 337.935V323.859M644.528 271.743L668.531 257.885L658.794 252.263L644.528 260.499M644.528 271.743V265.885M644.528 271.743V292.982M644.528 260.499V265.885M644.528 260.499V238.789M618.127 149.163L591.053 211.118M591.053 211.118L618.19 226.786M591.053 211.118V238.401M591.053 211.118L618.127 254.032M591.053 323.859L559.39 385.629M591.053 323.859V292.593M591.053 323.859L618.127 308.224M559.39 385.629V426.3M559.39 385.629L588.347 443.019M559.39 552.09V509.973M559.39 552.09L588.347 526.692M644.528 211.693V238.789M644.528 211.693L618.127 254.032M644.528 323.859V292.982M644.528 323.859L618.127 308.224M644.528 323.859L618.127 381.68M618.127 381.68V411.87V423.517M618.127 381.68L588.347 443.019M618.127 381.68L644.528 366.437M618.127 381.68L644.528 408.274M618.127 281.128L644.528 265.885M618.127 281.128V254.032M618.127 281.128V308.224M618.127 281.128L644.528 292.982M618.127 281.128L644.528 238.789M618.127 281.128L591.053 265.497M618.127 281.128L591.053 238.789M618.127 281.128L591.053 292.593M644.528 265.885L618.127 254.032M644.528 265.885L618.127 308.224M618.127 254.032L644.528 238.789M618.127 254.032L591.053 238.401M618.127 254.032L591.053 265.497M618.127 308.224L644.528 292.982M618.127 308.224L591.053 292.593M618.127 308.224L591.053 265.497M591.053 238.401V238.789M591.053 265.497V292.593M591.053 265.497V238.789M588.347 484.855V443.019M588.347 484.855V526.692M588.347 484.855L618.127 423.517M588.347 484.855L618.127 507.189M588.347 484.855L559.39 468.137M588.347 484.855L559.39 426.3M588.347 484.855L559.39 509.973M588.347 484.855L601.237 475.104L618.127 465.353M618.127 465.353V423.517M618.127 465.353V507.189M618.127 465.353L588.347 443.019M618.127 465.353L588.347 526.692M618.127 465.353L644.528 450.11M618.127 465.353L644.528 491.947M618.127 465.353L644.528 408.274M588.347 443.019L618.127 423.517M588.347 443.019L559.39 426.3M588.347 443.019L559.39 468.137M618.127 423.517L644.528 408.274M618.127 423.517L644.528 450.11M618.127 423.517L644.528 366.437M588.347 526.692L618.127 507.189M588.347 526.692L559.39 509.973M588.347 526.692L559.39 468.137M618.127 507.189L644.528 491.947M618.127 507.189L644.528 450.11M644.528 366.437V408.274M644.528 408.274V450.11M644.528 450.11V491.947M559.39 509.973V468.137M559.39 468.137V426.3"
+          stroke="url(#paint25_linear_4_2)"
+        />
+  </>
+);
+
+export default ScenePipesAndPulse;
